% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{addCellType}
\alias{addCellType}
\title{Adds a new cell type to a CellTypeHierarchy}
\usage{
addCellType(cth, cell_type_name, classify_func,
  parent_cell_type_name = "root")
}
\arguments{
\item{cth}{The CellTypeHierarchy object}

\item{cell_type_name}{The name of the new cell type. Can't already exist in cth}

\item{classify_func}{A function that returns true when a cell is of the new type}

\item{parent_cell_typ_name}{If this cell type is a subtype of another, provide its name here}
}
\value{
The updated CellTypeHierarchy
}
\description{
Adds a new cell type to a CellTypeHierarchy
}

