% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{assign_cell_state_helper}
\alias{assign_cell_state_helper}
\title{a function to assign pseudotime for the MST}
\usage{
assign_cell_state_helper(ordering_tree_res, curr_cell,
  visited_node = curr_cell)
}
\description{
a function to assign pseudotime for the MST
}

