% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{assignPseudotimeBranchPT}
\alias{assignPseudotimeBranchPT}
\title{a function to assign pseudotime and states based on the projected coordinates from the DDRTree algorithm, the stree matrix maybe used later
also: fix the bug when the scale_pseudotime can be used}
\usage{
assignPseudotimeBranchPT(cds, root_cell = NULL, scale_pseudotime = F,
  verbose = F)
}
\arguments{
\item{cds}{a matrix with N x N dimension}

\item{root_cell}{a dataframe used to generate new data for interpolation of time points}

\item{scale_pseudotime}{a matrix with N x N dimension}

\item{verbose}{a matrix with N x N dimension}
}
\value{
a cds object with the states and the branch assigned correctly
}
\description{
a function to assign pseudotime and states based on the projected coordinates from the DDRTree algorithm, the stree matrix maybe used later
also: fix the bug when the scale_pseudotime can be used
}

