% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{build_asym_kNN_graph}
\alias{build_asym_kNN_graph}
\title{function to run build asymmetric kNN graph}
\usage{
build_asym_kNN_graph(data, k = 20, dist_type = c("raw", "euclidean",
  "cosine"), return_graph = F)
}
\arguments{
\item{data}{low dimensional space used to perform graph clustering}

\item{k}{number of nearest neighbors used for Louvain clustering}

\item{return_graph}{whether or not to return the kNN graph instead of the asymmetric adjacency matrix}
}
\value{
Either a sparse asymmetric adjacent matrix or a corresponding directed weighted kNN graph
}
\description{
function to run build asymmetric kNN graph
}
