% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expr_models.R
\name{buildLineageBranchCellDataSet}
\alias{buildLineageBranchCellDataSet}
\title{Build a CellDataSet with appropriate duplication along two lineages}
\usage{
buildLineageBranchCellDataSet(cds, lineage_states = c(2, 3),
  lineage_labels = NULL, method = "fitting", stretch = FALSE,
  weighted = FALSE, gene_pairs = NULL, ...)
}
\description{
Build a CellDataSet with appropriate duplication along two lineages
}

