% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{cal_ncenter}
\alias{cal_ncenter}
\title{Function to decide a good number of centers for running DDRTree on big datasets}
\usage{
cal_ncenter(ncells, ncells_limit = 100)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}
}
\value{
a new cds containing only the genes used in reducing dimension. Expression values are reverse embedded.
}
\description{
Function to decide a good number of centers for running DDRTree on big datasets
}

