% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM.R
\name{calABCs}
\alias{calABCs}
\title{Calculate the area between TWO fitted lineage trajectories}
\usage{
calABCs(cds, trend_formula = "~sm.ns(Pseudotime, df = 3)*Lineage",
  trajectory_states = c(2, 3), relative_expr = TRUE, stretch = TRUE,
  pseudocount = 0, cores = 1, weighted = TRUE, verbose = F,
  min_expr = 0.5, integer_expression = FALSE, num = 5000,
  lineage_labels = NULL, ABC_method = c("integral"), ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{trend_formula}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{trajectory_states}{States corresponding to two branches}

\item{relative_expr}{a logic flag to determine whether or not the relative gene expression should be used}

\item{stretch}{a logic flag to determine whether or not each lineage should be stretched}

\item{pseudocount}{pseudo count added before fitting the spline curves}

\item{cores}{the number of cores to be used while testing each gene for differential expression}

\item{weighted}{a logic flag to determine whether or not we should use the navie logLikelihood weight scheme for the duplicated progenitor cells}

\item{verbose}{a logic flag to determine whether or not we should output detailed running information}

\item{min_expr}{the lower limit for the expressed gene}

\item{integer_expression}{the logic flag to determine whether or not the integer numbers are used for calculating the ABCs. Default is False.}

\item{num}{number of points on the fitted lineage trajectories used for calculating the ABCs. Default is 5000.}

\item{lineage_labels}{the name for each lineage, for example, AT1 or AT2}

\item{ABC_method}{the method used to calculate the ABC scores. Currently it only supports "integral". In future, it can be either one of "integral", "global_normalization", "local_normalization", "four_values", "ILRs".
"global_normalization" or "local_normalization" are similar measures between normalized by the global maximum or current maximum. 
"ILRs" is similar to calculate the Instant Log Ratio used in calILRs function}
}
\value{
a data frame containing the ABCs (Area under curves) score as the first column and other meta information from fData
}
\description{
This function is used to calculate the ABC score based on the the nature spline curves fitted for each lineage. ABC score is used to 
quantify the total magnitude of divergence between two lineages. By default, the ABC score is the area between two fitted spline curves. 
The ABC score can be used to rank gene divergence. When coupled with p-val calculated from the branchTest, it can be used to identify
potential major regulators for lineage bifurcation.
}

