% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{calculateExpressionThreshold}
\alias{calculateExpressionThreshold}
\title{Calculate a threshhold below which to exclude genes from ordering
based on the dispersion curve}
\usage{
calculateExpressionThreshold(cds, dispersionThresh = 0.1)
}
\description{
Calculate a threshhold below which to exclude genes from ordering
based on the dispersion curve
}

