% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{calculateLW}
\alias{calculateLW}
\title{Function to calculate the a neighbours list with spatial weights for the chosen coding scheme from a cell dataset object}
\usage{
calculateLW(cds, verbose = FALSE, k = 25, return_sparse_matrix = FALSE)
}
\description{
Function to calculate the a neighbours list with spatial weights for the chosen coding scheme from a cell dataset object
}
