% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{calculateMarkerSpecificity}
\alias{calculateMarkerSpecificity}
\title{Calculate each gene's specificity for each cell type}
\usage{
calculateMarkerSpecificity(cds, cth)
}
\value{
For a CellDataset with N genes, and a CellTypeHierarchy with k types,
returns a dataframe with N x k rows. Each row contains a gene and a specifity
score for one of the types.
}
\description{
Computes the Jensen-Shannon distance between the distribution of a gene's 
expression across cells and a hypothetical gene that is perfectly restricted
to each cell type. The Jensen-Shannon distance is an information theoretic
metric between two probability distributions. It is a widely accepted measure
of cell-type specificity. For a complete description see Cabili \emph{et. al},
Genes & Development (2011).
}

