% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calibrate_mode_analytical}
\alias{calibrate_mode_analytical}
\title{Function used to calibrate the mode, m and c as well as the total mRNA using analytical formula:}
\usage{
calibrate_mode_analytical(ind, tpm_distribution, t_estimate, ladder,
  total_ladder_transcripts, total_mRNA, capture_rate)
}
\description{
Function used to calibrate the mode, m and c as well as the total mRNA using analytical formula:
}

