% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calibrate_per_cell_total_proposal}
\alias{calibrate_per_cell_total_proposal}
\title{Calibrate_per_cell_total_proposal}
\usage{
calibrate_per_cell_total_proposal(relative_exprs_matrix, t_estimate,
  expected_capture_rate)
}
\arguments{
\item{relative_exprs_matrix}{The matrix of relative TPM expression values}

\item{t_estimate}{the TPM value that corresponds to 1 cDNA copy per cell}

\item{expected_capture_rate}{The fraction of mRNAs captured as cDNAs}
}
\description{
Calibrate_per_cell_total_proposal
}

