% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\docType{methods}
\name{cellPairwiseDistances}
\alias{cellPairwiseDistances}
\title{Retrieves a matrix capturing distances between each cell in the reduced-dimensionality space}
\usage{
cellPairwiseDistances(cds)
}
\arguments{
\item{cds}{expression data matrix for an experiment}
}
\value{
A square, symmetric matrix containing the distances between each cell in the reduced-dimensionality space.
}
\description{
Retrieves a matrix capturing distances between each cell in the reduced-dimensionality space
}
\examples{
\dontrun{
D <- cellPairwiseDistances(HSMM)
}
}

