% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellTypeHierarchy.R
\docType{class}
\name{CellTypeHierarchy}
\alias{CellTypeHierarchy}
\alias{CellTypeHierarchy-class}
\title{The CellTypeHierarchy class}
\description{
Classifies cells according to a hierarchy of types.
}
\details{
Classifies cells according to a hierarchy of types via user-defined
gating functions.
}
\section{Slots}{
 
 \describe{
   \item{\code{classificationTree}:}{Object of class \code{"igraph"}}
 }
}

