% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{classifyCellClusters}
\alias{classifyCellClusters}
\title{Impute cell type classifications using clustering results}
\usage{
classifyCellClusters(cds, cth, frequency_thresh = NULL,
  enrichment_thresh = NULL, ...)
}
