% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{classifyCells}
\alias{classifyCells}
\title{Classify cells according to a set of markers}
\usage{
classifyCells(cds, cth, frequency_thresh, cores = 1, ...)
}
\description{
Classify cells according to a set of markers
}

