% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{classifyCellsHelperCellGlmNet}
\alias{classifyCellsHelperCellGlmNet}
\title{classifyCellsHelperCellGlmNet}
\usage{
classifyCellsHelperCellGlmNet(cds, cth)
}
\description{
Classify cells using glmnet
}
