% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterCells}
\alias{clusterCells}
\title{Cluster cells into a specified number of groups.}
\usage{
clusterCells(cds, num_clusters, cell_type_hierarchy = NULL,
  frequency_thresh = NULL, enrichment_thresh = NULL,
  clustering_genes = NULL, max_components = 10,
  residualModelFormulaStr = NULL, param.gamma = 100, verbose = F, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{num_clusters}{number of desired cell clusters}

\item{cell_type_hierarchy}{the CellTypeHierarchy that divides the cells from cds into different types of cells}

\item{frequency_thresh}{When a CellTypeHierarchy is provided, cluster cells will impute cell types in clusters that are composed of at least this much of exactly one cell type.}

\item{clustering_genes}{a vector of genes used to differentiate between the cell types in the CellTypeHierarchy}

\item{max_components}{number of dimensions to project the data into via \code{\link{reduceDimension}()}}

\item{residualModelFormulaStr}{A model formula specifying the effects to subtract from the data before clustering.}

\item{param.gamma}{gamma parameter for DDRTree}

\item{verbose}{Verbose parameter for DDRTree}

\item{...}{Additional arguments passed to \code{\link{reduceDimension}()}}
}
\value{
an updated CellDataSet object, in which phenoData contains values for Cluster for each cell
}
\description{
Unsupervised clustering of cells is a common step in many single-cell expression
workflows. In an experiment containing a mixture of cell types, each cluster might
correspond to a different cell type. This method takes a CellDataSet as input
along with a requested number of clusters, clusters them with an unsupervised 
algorithm, and then returns the CellDataSet with the cluster assignments stored in
the pData table.
}

