% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{clusterCells_Density_Peak}
\alias{clusterCells_Density_Peak}
\title{Cluster cells into a specified number of groups based on .}
\usage{
clusterCells_Density_Peak(cds, skip_rho_sigma = F, variance_explained = 0.8,
  inspect_rho_sigma = F, rho_threshold = NULL, delta_threshold = NULL,
  peaks = NULL, cell_type_hierarchy = NULL, frequency_thresh = 0.1,
  clustering_genes = NULL, verbose = F, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{skip_rho_sigma}{A logic flag to determine whether or not you want to skip the calculation of rho / sigma}

\item{variance_explained}{Variance explained by the PCA components, used to select the component numbers for tSNE}

\item{inspect_rho_sigma}{A logical flag to determine whether or not you want to interactively select the rho and sigma for assigning up clusters}

\item{rho_threshold}{The threshold of local density (rho) used to select the density peaks}

\item{delta_threshold}{The threshold of local distance (delta) used to select the density peaks}

\item{peaks}{A numeric vector indicates the index of density peaks used for clustering. This vector should be retrieved from the decision plot with caution. No checking involved.
will automatically calculated based on the top num_cluster product of rho and sigma.}

\item{frequency_thresh}{When a CellTypeHierarchy is provided, cluster cells will impute cell types in clusters that are composed of at least this much of exactly one cell type.}

\item{verbose}{Verbose parameter for DDRTree}

\item{...}{Additional arguments passed to \code{\link{densityClust}()}}
}
\value{
an updated CellDataSet object, in which phenoData contains values for Cluster for each cell
}
\description{
Unsupervised clustering of cells is a common step in many single-cell expression
workflows. In an experiment containing a mixture of cell types, each cluster might
correspond to a different cell type. This method takes a CellDataSet as input
along with a requested number of clusters, clusters them with an unsupervised 
algorithm, and then returns the CellDataSet with the cluster assignments stored in
the pData table. When number of clusters is set to NULL (num_clusters = NULL), 
the decision plot as introduced in the above citation will be plotted and the 
users are required to click on the decision plot to select the rho and delta 
to determine the number of clusters to cluster.
}
\references{
Rodriguez, A., & Laio, A. (2014). Clustering by fast search and find of density peaks. Science, 344(6191), 1492-1496. doi:10.1126/science.1242072
}

