% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{coefficientTable}
\alias{coefficientTable}
\title{Extracts a table of coefficients from a tibble containing model objects}
\usage{
coefficientTable(model_tbl)
}
\description{
Extracts a table of coefficients from a tibble containing model objects
}
