% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{compareModels}
\alias{compareModels}
\title{Compare model fits}
\usage{
compareModels(full_models, reduced_models)
}
\arguments{
\item{full_models}{a list of models, e.g. as returned by fitModels(), forming the numerators of the L.R.Ts.}

\item{reduced_models}{a list of models, e.g. as returned by fitModels(), forming the denominators of the L.R.Ts.}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
Performs likelihood ratio tests on nested vector generalized additive models
}
