% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{DDRTree}
\alias{DDRTree}
\title{Perform DDRTree construction}
\usage{
DDRTree(X, params, verbose = F)
}
\arguments{
\item{X}{a matrix with D x N dimension which is needed to perform DDRTree construction}

\item{params}{a list with the following parameters: 
maxIter : maximum iterations
eps     : relative objective difference
dim     : reduced dimension
lambda  : regularization parameter for inverse graph embedding
sigma   : bandwidth parameter
gamma   : regularization parameter for k-means}
}
\value{
a list with W, Z, stree, Y, history
}
\description{
Perform DDRTree construction
}

