% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM.R
\name{detectBifurcationPoint}
\alias{detectBifurcationPoint}
\title{Detect the maturation time point where the gene expression starts to diverge}
\usage{
detectBifurcationPoint(str_log_df = NULL, ILRs_threshold = 0.1,
  detect_all = T, cds = cds, Lineage = "Lineage", branch_point = NULL,
  lineage_states = c(2, 3), stretch = T, cores = detectCores(),
  trend_formula = "~sm.ns(Pseudotime, df = 3)", ILRs_limit = 3,
  relative_expr = TRUE, weighted = FALSE, label_by_short_name = TRUE,
  useVST = FALSE, round_exprs = FALSE, pseudocount = 0,
  output_type = "all", return_cross_point = T,
  file = "bifurcation_heatmap", verbose = FALSE, ...)
}
\arguments{
\item{str_log_df}{the ILRs dataframe calculated from calILRs function. If this data.frame is provided, all the following parameters are ignored. Note that we need to only use the ILRs after the bifurcation point if we duplicated the progenitor cell state.}

\item{ILRs_threshold}{the ILR value used to determine the earliest divergence time point}

\item{detect_all}{a logic flag to determine whether or not genes without ILRs pass the threshold will still report a bifurcation point}

\item{cds}{CellDataSet for the experiment}

\item{Lineage}{The column in pData used for calculating the ILRs (If not equal to "Lineage", a warning will report)}

\item{branch_point}{the id for the branch point choosed to select cell branches and progenitors}

\item{lineage_states}{The states for two branching lineages}

\item{stretch}{a logic flag to determine whether or not each lineage should be stretched}

\item{cores}{Number of cores when fitting the spline curves}

\item{trend_formula}{the model formula to be used for fitting the expression trend over pseudotime}

\item{ILRs_limit}{the minimum Instant Log Ratio used to make the heatmap plot}

\item{relative_expr}{A logic flag to determine whether or not the relative expressed should be used when we fitting the spline curves}

\item{weighted}{A logic flag to determine whether or not we should use the navie logLikelihood weight scheme for the duplicated progenitor cells}

\item{label_by_short_name}{label the rows of the returned matrix by gene_short_name (TRUE) or feature id (FALSE)}

\item{useVST}{A logic flag to determine whether or not the Variance Stablization Transformation should be used to stablize the gene expression.
When VST is used, the difference between two lineages are used instead of the log-ratio.}

\item{round_exprs}{A logic flag to determine whether or not the expression value should be rounded into integer}

\item{pseudocount}{pseudo count added before fitting the spline curves}

\item{output_type}{A character either of "all" or "after_bifurcation". If "after_bifurcation" is used, only the time points after the bifurcation point will be selected. Note that, if Lineage is set to "Lineage", we will only use "after_bifurcation" since we duplicated the progenitor cells and the bifurcation should only happen after the largest mature level from the progenitor cells}

\item{return_cross_point}{A logic flag to determine whether or not only return the cross point}

\item{file}{the name for storing the data. Since the calculation of the Instant Log Ratio is very time consuming, so by default the result will be stored}
}
\value{
a vector containing the time for the bifurcation point with gene names for each value
}
\description{
This function is used to determine the bifurcation point for the gene expression between two distinct biological processes.
For processes we can not distinguish between lineages (or phenotype groups, like knockout VS un-knockout), this function will 
only detect bifurcation points after the inferenced bifurcatioin from the PQ-tree. The
}

