% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detectGenes}
\alias{detectGenes}
\title{Detects genes above minimum threshold.}
\usage{
detectGenes(cds, min_expr = NULL)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{min_expr}{the expression threshold}
}
\value{
an updated CellDataSet object
}
\description{
Sets the global expression detection threshold to be used with this CellDataSet.
Counts how many cells each feature in a CellDataSet object that are detectably expressed 
above a minimum threshold. Also counts the number of genes above this threshold are 
detectable in each cell.
}
\examples{
\dontrun{
HSMM <- detectGenes(HSMM, min_expr=0.1)
}
}
