% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{diff_test_helper}
\alias{diff_test_helper}
\title{Helper function for parallel differential expression testing}
\usage{
diff_test_helper(x, fullModelFormulaStr, reducedModelFormulaStr,
  expressionFamily, relative_expr, weights, disp_func = NULL,
  pseudocount = 0, verbose = FALSE)
}
\arguments{
\item{relative_expr}{Whether to transform expression into relative values}
}
\description{
Helper function for parallel differential expression testing
}

