% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_selection.R
\name{downsampleCDS}
\alias{downsampleCDS}
\title{Downsample datasets with landmark algorithm to ensure full sampling of the entire data space}
\usage{
downsampleCDS(cds, landmark_num)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}

\item{landmark_num}{number of landmark cells to retrieve}
}
\value{
a new cds containing only the cells selected by the landmark selection algorithm
}
\description{
Downsample datasets with landmark algorithm to ensure full sampling of the entire data space
}
