% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{extract_ddrtree_ordering}
\alias{extract_ddrtree_ordering}
\title{a function to assign pseudotime for the MST}
\usage{
extract_ddrtree_ordering(cds, root_cell, verbose = T)
}
\description{
a function to assign pseudotime for the MST
}

