% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{extract_good_branched_ordering}
\alias{extract_good_branched_ordering}
\title{Extract a linear ordering of cells from a PQ tree}
\usage{
extract_good_branched_ordering(orig_pq_tree, curr_node, dist_matrix,
  num_branches, reverse_main_path = FALSE)
}
\description{
Extract a linear ordering of cells from a PQ tree
}

