% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{FDL}
\alias{FDL}
\title{Run improved force directed layout for cells in low dimensional space.}
\usage{
FDL(cds, separate_group = TRUE, method = c("drl", "fr", "kk"),
  merge_coords_method = c("procrutes", "dla"), start.temp = NULL, color_by,
  markers = NULL, cell_size = 1, cell_link_size = 1, k = 25,
  cell_num_threshold = 100, verbose = FALSE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{separate_group}{Whether or not to separate participation groups and apply FDL in each partition or directly select equal number of representatives in each louvain groups}

\item{method}{The force directed layout method to use}

\item{merge_coords_method}{The method used to patch different disconnected component into a single graph}

\item{start.temp}{argument passed into layout_with_fr function}

\item{k}{Number of nearest neighbors used in calculating the force direct layout}

\item{cell_num_threshold}{The minimum number of cells to be ignore for each partition component}

\item{verbose}{Wheter to print all running details}

\item{...}{additional arguments passed to functions (louvain_clustering) called by this function.}
}
\value{
a ggplot2 plot object or a list of all computated information from this function
}
\description{
Run improved force directed layout for cells in low dimensional space.
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
HSMM <- reduceDimension(HSMM, reduction_method = 'UMAP')
HSMM <- smoothEmbedding(HSMM, smooth_method = 'FDL', verbose = T)
}
}
