% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_cluster_markers}
\alias{find_cluster_markers}
\title{Find marker genes for each group of cells}
\usage{
find_cluster_markers(cds, spatial_res, group_by = "Cluster",
  qval_threshold = 0.05, morans_I_threshold = 0.25, lower_threshold = 0,
  pseudocount = 1, top_n_by_group = NULL, verbose = FALSE, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{spatial_res}{the result returned from spatialDifferentialTest}

\item{group_by}{a column in the pData specifying the groups for calculating the specifities. By default it is Cluster}

\item{qval_threshold}{The q-value threshold for genes to be selected}

\item{morans_I_threshold}{The lowest Morans' I threshold for selecting genes}

\item{lower_threshold}{The lowest gene expression threshold for genes to be considered as expressed}

\item{pseudocount}{Pseduo-count added to gene expression before calculating the log}

\item{top_n_by_group}{Select top_n_by_group from each group based on the specificity}

\item{verbose}{Whether to show VGAM errors and warnings. Only valid for cores = 1.}

\item{...}{Additional arguments passed to function}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
The common practice to identify cluster specific genes only involves testing differentially expressed genes across one cluster of cells
with another cluster of cells (or alternatively all other cells). This often involves a lot gene filtering step (either filter based on 
number of cells expresed in a cluster, difference of gene expression) and heavy computation. Moreover, it may ignore that fact that some 
genes show certain gradient of gene expression and may span multiple clusters. With the Moran's I test, we can avoid those limitations 
and identify all genes significantly spatially correlated all in once very efficient. To categorize genes into cluster specific genes, 
we simply calculate the specificity of each gene to each cell cluster.
In Monocle 3, we incoporated the find_cluster_markers to find marker genes for each group of cells. In order to calculate the specificity 
of each gene to each cluster, we first create a perfect specific distribution ( Q ) of each gene by assigning probability 1 to one particular 
cluster while all 0 to any other clusters. Then we calculate the percentage of cells expressed in each cluster and convert this percentage 
to a probability distribution ( P ) by dividing the sum of the percentages across all clusters. Then we calculate the specificity score 
as 1 - Jensen-Shannon distance between this distribution to the perfect specific distribution.
}
\seealso{
\code{\link[monocle]{principalGraphTest}}
}
