% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/expr_models.R
\name{fit_model_helper}
\alias{fit_model_helper}
\title{Helper function for parallel VGAM fitting}
\usage{
fit_model_helper(x, modelFormulaStr, expressionFamily, relative_expr)
}
\description{
Helper function for parallel VGAM fitting
}

