% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{fpkm2abs}
\alias{fpkm2abs}
\title{Transform FPKM matrix to absolute transcript matrix based on the decomposed linear regression parameters from most abundant isoform fpkm value.}
\usage{
fpkm2abs(fpkm_matrix, t_estimate = estimate_t(fpkm_matrix),
  total_fragment = 1500000, alpha_v = 1, m = -3.652201, c = 2.263576,
  global_scaling = FALSE, return_all = FALSE, num_cores = 1)
}
\arguments{
\item{fpkm_matrix}{an matrix of fpkm for single cell RNA-seq values with each row and column representing genes/isoforms and cells. Row and column names should be included}

\item{t_estimate}{an vector for the estimated most abundant FPKM value of isoform for a single cell. Estimators based on gene FPKM can also give good approximation but estimators
based on isoform FPKM will give better results in general}

\item{global_scaling}{parameter for globaling scaling. Not perform global scaling by default (global_scaling = 1)}

\item{return_all}{parameter for the intended return results. If setting TRUE, matrix of k^*, b^* and vector of global_scaling as well the transformed absolute cds will be returned
in a list format}

\item{total_fragment:}{vector of total fragment sequenced for each cell, the element should matched with fpkm_matrix columns}

\item{alpha_v:}{vector of most abundant transcripts number estimates in each cell, which are dependent on the total_fragment}
}
\value{
an matrix of absolute count for isoforms or genes after the transformation. For more details on other output, please refer to detail
}
\description{
Transform FPKM matrix to absolute transcript matrix based on the decomposed linear regression parameters from most abundant isoform fpkm value.
}
\details{
This function takes a FPKM matrix and a vector of estimated most abundant FPKM value from the isoform FPKM matrix and transform it into absolute transcript number.
It is based on the fact that most isoforms of gene in a single cell only express one copy so that the most abundant isoform FPKM (t^*) will corresponding to 1 copy transcript. The
function takes the the vector t^* and then decomposes it into two parameters vectors k^* and b^* (corresponding to each cell) which correspond to the slope and intercept when
we perform the robust linear regression for the spikein data. This decomposition is based on an observed relationship between k and b in terms of b = -3.652201 k + 2.263576. The
function will then apply a linear transformation to convert the FPKM to estimated absolute transcript counts based on the the k^* and b^*. The function can also apply a global
adjustment if setting global_scaling = TRUE. The k*, b* parameters vectors and the global scaling factor can be output in a list format (together with norm_cds) if setting return_all
== TRUE
}
\examples{
\dontrun{
HSMM_fpkm_matrix <- exprs(HSMM)
HSMM_abs_matrix <- fpkm2abs(HSMM_fpkm_matrix, t_estimate = estimate_t(fpkm_matrix))
}
accept an log fpkm matrix and transform it into the absolute transcript matrix

row and column names of the matrix are genes and cells respectively
}

