% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_correct_root_state}
\alias{get_correct_root_state}
\title{A helper function to identify the root principal points}
\usage{
get_correct_root_state(cds, cell_phenotype, root_type)
}
\description{
A helper function to identify the root principal points
}
