% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{get_mc_list}
\alias{get_mc_list}
\title{Return the slopes and intercepts matrix for the relationship between regression parameters Ks, Bs in all cells at different concentration detection limit.
The slopes/intercepts for different concentration can be obtained through the row names}
\usage{
get_mc_list()
}
\description{
Return the slopes and intercepts matrix for the relationship between regression parameters Ks, Bs in all cells at different concentration detection limit.
The slopes/intercepts for different concentration can be obtained through the row names
}

