% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{ica_helper}
\alias{ica_helper}
\title{Run the fastICA algorithm on a numeric matrix.}
\usage{
ica_helper(X, n.comp, alg.typ = c("parallel", "deflation"),
  fun = c("logcosh", "exp"), alpha = 1, row.norm = TRUE, maxit = 200,
  tol = 1e-04, verbose = FALSE, w.init = NULL, use_irlba = TRUE)
}
\description{
Run the fastICA algorithm on a numeric matrix.
}

