% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{JSdistVec}
\alias{JSdistVec}
\title{Calculate the Jessen-Shannon distance for two probability distribution}
\usage{
JSdistVec(p, q)
}
\description{
Calculate the Jessen-Shannon distance for two probability distribution
}

