% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landmark_selection.R
\name{landmark_selection}
\alias{landmark_selection}
\title{Downsample datasets with landmark algorithm to ensure full sampling of the entire data space}
\usage{
landmark_selection(cds, landmark_num)
}
\arguments{
\item{cds}{a cell dataset for landmark selection}

\item{landmark_num}{number of landmark cells to retrieve}
}
\value{
a list of three named vectors with the same number of cells from the cds: assign (which landmark the current cell closest to), 
dist (distance of current cell to its closest landmark) and flag (whether or not the current cell is a landmark, 1: yes, 0: no)
}
\description{
Downsample datasets with landmark algorithm to ensure full sampling of the entire data space
}
