% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_HSMM_markers}
\alias{load_HSMM_markers}
\title{Return a CellDataSet of classic muscle genes.}
\usage{
load_HSMM_markers()
}
\value{
A CellDataSet object
}
\description{
Return a CellDataSet of classic muscle genes.
}
