% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/louvain.R
\encoding{UTF-8}
\name{louvain_R}
\alias{louvain_R}
\title{Uniform Manifold Approximation and Projection}
\usage{
louvain_R(X, python_home = system("which python", intern = TRUE),
  partition_method = "CPMVertexPartition", initial_membership = NULL,
  weights = NULL, res = 0.6, node_sizes = NULL, verbose = FALSE,
  return_all = FALSE)
}
\arguments{
\item{X}{the dataset upon which to perform umap dimension reduction}

\item{python_home}{The python home directory where umap is installed}

\item{partition_method}{character - either the default "CPMVertexPartition" or  "RBConfigurationVertexPartition" / "RBERVertexPartition".}

\item{initial_membership}{(list of int) – Initial membership for the partition. If None then defaults to a singleton partition.}

\item{weights}{(list of double, or edge attribute) – Weights of edges. Can be either an iterable or an edge attribute.}

\item{res}{(double) – Resolution parameter.}

\item{node_sizes}{(list of int, or vertex attribute) – Sizes of nodes are necessary to know the size of communities in aggregate graphs. Usually this is set to 1 for all nodes, but in specific cases this could be changed.}

\item{verbose}{bool (optional, default False)}

\item{return_all}{Whether to return all slots after louvain 
Controls verbosity of logging.}
}
\value{
The cluster id if return_all set to be FALSE, otherwise all slots from the louvain function
}
\description{
Finds a low dimensional embedding of the data that approximates an underlying manifold.
This functions relies on the python implementation of UMAP (https://github.com/lmcinnes/umap). 
The original publication of UMAP can be found here: 
https://arxiv.org/abs/1802.03426 (McInnes, L, Healy, J, UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction, ArXiv e-prints 1802.03426, 2018).
A useful notebook (written in python) to check for the effects of each parameter in UMAP can be found here: https://nbviewer.jupyter.org/github/CrakeNotSnowman/umapNotebooks/blob/master/UMAP%20Usage.ipynb.
}
