% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{make_heatmap}
\alias{make_heatmap}
\title{Make heatmap}
\usage{
make_heatmap(dat.all = logfc_df, emsemble_ids = quake_id,
  file = "heatmap.pdf", ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{rowgenes}{Gene ids or short names to be arrayed on the vertical axis.}

\item{colgenes}{Gene ids or short names to be arrayed on the horizontal axis}

\item{relative_expr}{Whether to transform expression into relative values}

\item{min_expr}{The minimum level of expression to show in the plot}
}
\value{
a ggplot2 plot object
}
\description{
Make heatmap
}

