% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{makeprobs}
\alias{makeprobs}
\title{Calculate the probability matrix for a relative abundance matrix}
\usage{
makeprobs(a)
}
\description{
Calculate the probability matrix for a relative abundance matrix
}

