% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{markerDiffTable}
\alias{markerDiffTable}
\title{Test genes for cell type-dependent expression}
\usage{
markerDiffTable(cds, cth, residualModelFormulaStr = "~1", max_cells = NULL,
  cores = 1)
}
\arguments{
\item{cds}{A CellDataSet object containing cells to classify}

\item{cth}{The CellTypeHierarchy object to use for classification}

\item{residualModelFormulaStr}{A model formula string specify effects you
want to exclude when testing for cell type dependent expression}

\item{cores}{The number of cores to use when testing}
}
\value{
A table of differential expression test results
}
\description{
Test genes for cell type-dependent expression
}

