% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\name{minSpanningTree<-}
\alias{minSpanningTree<-}
\title{Sets the minimum spanning tree used by Monocle during cell ordering. Not intended to be called directly.}
\usage{
minSpanningTree(cds) <- value
}
\arguments{
\item{cds}{A CellDataSet object.}

\item{value}{an igraph object describing the minimum spanning tree.}
}
\value{
An updated CellDataSet object
}
\description{
Sets the minimum spanning tree used by Monocle during cell ordering. Not intended to be called directly.
}
\examples{
\dontrun{
cds <- minSpanningTree(T)
}
}

