% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{opt_norm_t}
\alias{opt_norm_t}
\title{Recover the absolute transcript counts based on m,c and t estimate for a single cell (Used in the optimization function)}
\usage{
opt_norm_t(t, fpkm, m, c, expr_thresh = 0.1, pseudocnt = NULL,
  return_norm = FALSE)
}
\description{
Recover the absolute transcript counts based on m,c and t estimate for a single cell (Used in the optimization function)
}

