% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/order_cells.R
\name{order_p_node}
\alias{order_p_node}
\title{Return an ordering for a P node in the PQ tree}
\usage{
order_p_node(q_level_list, dist_matrix)
}
\description{
Return an ordering for a P node in the PQ tree
}

