% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{orderCells}
\alias{orderCells}
\title{Orders cells according to progress through a learned biological process.}
\usage{
orderCells(cds, root_state = NULL, num_paths = NULL, reverse = NULL)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{root_state}{The state to use as the root of the trajectory. 
You must already have called orderCells() once to use this argument.}

\item{num_paths}{the number of end-point cell states to allow in the biological process.}

\item{reverse}{whether to reverse the beginning and end points of the learned biological process.}
}
\value{
an updated CellDataSet object, in which phenoData contains values for State and Pseudotime for each cell
}
\description{
Orders cells according to progress through a learned biological process.
}

