% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{patchEmbedding}
\alias{patchEmbedding}
\title{Run embedding smoothing techniques (drl, PSL and SSE) for each isolated group and then put different embedding in the same coordinate system}
\usage{
patchEmbedding(cds, max_components = 2, do_partition = FALSE,
  use_pca = FALSE, method = c("PSL", "drl", "fr", "kk", "SSE"),
  merge_coords_method = c("dla", "procrutes"), start.temp = NULL,
  k = 20, landmark_num = 2000, cell_num_threshold = 0,
  verbose = FALSE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{max_components}{the dimensionality of the reduced space}

\item{do_partition}{Whether or not to separate participation groups and apply FDL in each partition or directly select equal number of representatives in each louvain groups}

\item{use_pca}{Whether or not to cluster cells based on top PCA component. Default to be FALSE.}

\item{method}{The embedding smoothing technique to use, including force directed layout (which includes drl, fr, kk three methods), our new method PSL and SSE}

\item{merge_coords_method}{The method used to patch different smoothed embedding from disconnected component into a coordinate system}

\item{start.temp}{argument passed into layout_with_fr function}

\item{k}{Number of nearest neighbors used in calculating the force direct layout}

\item{landmark_num}{Number of landmark used for downsampling the data. Currently landmarks are defined as the medoids from kmean clustering.}

\item{cell_num_threshold}{The minimum number of cells in each partition component required for running embedding smooth}

\item{verbose}{Wheter to print all running details}

\item{...}{additional arguments passed to functions (louvain_clustering) called by this function.}
}
\value{
a cds with smoothed embedding for each disconnected component learned (stored in reduceDimS, reduceDimK) and the corresponding graph stored in minSpanningTree
}
\description{
Run embedding smoothing techniques (drl, PSL and SSE) for each isolated group and then put different embedding in the same coordinate system
}
\references{
PSL: Li Wang, Qi Mao (2018). Probabilistic Dimensionality Reduction via Structure Learning. IEEE Transactions on Pattern Analysis and Machine Intelligence

SSE: Li Wang, Qi Mao, Ivor W. Tsang (2017). Latent Smooth Skeleton Embedding. Proceedings of the 31th AAAI Conference on Artificial Intelligence. 2017.
}
