% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DDRTree.R
\name{pca_projection}
\alias{pca_projection}
\title{perform PCA projection
solve the problem size(C) = NxN, size(W) = NxL
max_W trace( W' C W ) : W' W = I}
\usage{
pca_projection(C, L)
}
\arguments{
\item{C}{a matrix with N x N dimension}

\item{L}{a dataframe used to generate new data for interpolation of time points}
}
\value{
a matrix (W) with N x L dimension
}
\description{
perform PCA projection
solve the problem size(C) = NxN, size(W) = NxL
max_W trace( W' C W ) : W' W = I
}

