% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_3d_cell_trajectory}
\alias{plot_3d_cell_trajectory}
\title{Plot a dataset and trajectory in 3 dimensions}
\usage{
plot_3d_cell_trajectory(cds, color_by = NULL, markers = NULL,
  markers_linear = FALSE, webGL_filename = NULL, movie_filename = NULL,
  show_backbone = TRUE, scale_expr = TRUE, palette = NULL,
  useNULL_GLdev = FALSE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{markers_linear}{a boolean used to indicate whether you want to scale the markers logarithimically or linearly}

\item{webGL_filename}{the name of a file to which you'd to write a webGL file containing the plot}

\item{movie_filename}{the name of a file to which you'd to write an animated GIF containing the plot}

\item{show_backbone}{whether to show the principal graph used to order the cells}

\item{scale_expr}{whether to tranform the log expression values to z scores}

\item{palette}{the color palette used for plotting,}

\item{useNULL_GLdev}{if TRUE, don't show the plot on the screen (to be used with webGL or movie output)}
}
\value{
a ggplot2 plot object
}
\description{
Plot a dataset and trajectory in 3 dimensions
}
\examples{
\dontrun{
plot_3d_cell_trajectory(cds, markers=c("Rbfox3, Neurod1", "Sox2"))
}

}
