% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_avgerage_markers_cluster}
\alias{plot_avgerage_markers_cluster}
\title{A function to show the average expression pattern for gene clusters on low dimensional embedding}
\usage{
plot_avgerage_markers_cluster(cds, markers, return_all = FALSE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{markers}{A vector specifying the significant genes you want to cluster}

\item{return_all}{A logic flag to determine whether or not we should also return the clustered gene expression}

\item{...}{additional arguments passed into the function. Not used}
}
\value{
a ggplot2 plot object or a list with the plot object and a "cellDataset" storing the average expression for each cluster of genes across cells
}
\description{
A function to show the average expression pattern for gene clusters on low dimensional embedding
}
