% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cell_fdl}
\alias{plot_cell_fdl}
\title{Plots force directed layout of cells .}
\usage{
plot_cell_fdl(cds, layout = NULL, color_by = "Pseudotime", markers = NULL,
  cell_size = 0.5, cell_link_size = 1)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{layout}{the layout function used to generate the coordinates of cells}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{cell_size}{The size of the point for each cell}

\item{cell_link_size}{the maximal width of the link of edges (width of other edges are scaled correspondingly)}

\item{...}{additional arguments passed into the scale_color_viridis function}
}
\value{
a ggplot2 plot object
}
\description{
Plots force directed layout of cells .
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
HSMM <- reduceDimension(HSMM, reduction_method = 'UMAP')
plot_cell_fdl(HSMM)
plot_cell_fdl(HSMM, color_by="Pseudotime")
plot_cell_fdl(HSMM, markers="MYH3")
}
}
