% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_graph}
\alias{plot_cluster_graph}
\title{Plots a graph of louvain cluster, layouted with layout_component function, which can be used to remove outlier clusters of cells}
\usage{
plot_cluster_graph(cds)
}
\arguments{
\item{cds}{CellDataSet for the experiment}
}
\value{
a ggplot2 plot object
}
\description{
Plots a graph of louvain cluster, layouted with layout_component function, which can be used to remove outlier clusters of cells
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
HSMM <- reduceDimension(HSMM, method = 'UMAP')
plot_cluster_graph(HSMM)
}

}
