% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genes_branched_heatmap}
\alias{plot_genes_branched_heatmap}
\title{Create a heatmap to demonstrate the bifurcation of gene expression along two lineages}
\usage{
plot_genes_branched_heatmap(cds_subset, num_clusters = 6,
  lineage_states = c(2, 3), branch_point = NULL,
  lineage_labels = c("Cell fate 1", "Cell fate 2"), stretch = T,
  scaling = T, norm_method = "vstExprs", use_fitting_curves = T,
  dist_method = NULL, hclust_method = "ward", hmcols = NULL,
  Lineage_colors = c("#979797", "#F05662", "#7990C8"),
  trend_formula = "~sm.ns(Pseudotime, df=3) * Lineage",
  add_annotation_row = NULL, add_annotation_col = NULL, show_rownames = F,
  use_gene_short_name = F, return_heatmap = FALSE, cores = 1)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment (normally only the branching genes detected with branchTest)}

\item{num_clusters}{Number of clusters for the heatmap of branch genes}

\item{lineage_labels}{The label for the lineages (first (second) lineage corresponding to state 2 (3))}

\item{norm_method}{Either "vstExprs" or "log". Determine how to transform expression values prior to rendering}

\item{dist_method}{The method to calculate distance for each gene used in the hirearchical clustering, any one of them: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". (This option is not in function for now)}

\item{hclust_method}{The method to perform hirearchical clustering, any one of them: ward", "single", "complete", "average", "mcquitty", "median", "centroid"}

\item{hmcols}{The color scheme for drawing the heatmap}

\item{Lineage_colors}{The color for the progenitors and two lineages}

\item{return_heatmap}{Whether to return the pheatmap object to the user.}

\item{cores}{Number of cores to run this function}

\item{cores}{The number of cores to use when generating smoothing curves for rendering.}
}
\value{
A list of heatmap_matrix (expression matrix for the lineage committment), ph (pheatmap heatmap object),
annotation_row (annotation data.frame for the row), annotation_col (annotation data.frame for the column).
}
\description{
Create a heatmap to demonstrate the bifurcation of gene expression along two lineages
}

