% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genes_branched_pseudotime}
\alias{plot_genes_branched_pseudotime}
\title{Plot the branch genes in pseduotime with separate lineage curves
TO DO: add the confidence interval, add p-val, fix error when we cannot generate fitting for the data}
\usage{
plot_genes_branched_pseudotime(cds, lineage_states = c(2, 3),
  lineage_labels = NULL, method = "fitting", stretch = TRUE,
  min_expr = NULL, cell_size = 0.75, nrow = NULL, ncol = 1,
  panel_order = NULL, color_by = "State",
  trend_formula = "~ sm.ns(Pseudotime, df=3) * Lineage",
  label_by_short_name = TRUE, weighted = TRUE, add_ABC = FALSE,
  normalize = TRUE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{nrow}{number of columns used to layout the faceted cluster panels}

\item{ncol}{number of columns used to layout the faceted cluster panels}

\item{row_samples}{how many genes to randomly select from the data}
}
\value{
a ggplot2 plot object
}
\description{
Plot the branch genes in pseduotime with separate lineage curves
TO DO: add the confidence interval, add p-val, fix error when we cannot generate fitting for the data
}
\examples{
\dontrun{
full_model_fits <- fitModel(HSMM_filtered[sample(nrow(fData(HSMM_filtered)), 100),],  modelFormulaStr="~VGAM::bs(Pseudotime)")
expression_curve_matrix <- responseMatrix(full_model_fits)
clusters <- clusterGenes(expression_curve_matrix, k=4)
plot_clusters(HSMM_filtered[ordering_genes,], clusters)
}
}

