% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genes_heatmap}
\alias{plot_genes_heatmap}
\title{Plots a pseudotime-ordered, row-centered heatmap}
\usage{
plot_genes_heatmap(cds, rescaling = "row", clustering = "row",
  labCol = FALSE, labRow = TRUE, logMode = TRUE, pseudocount = 0.1,
  use_vst = TRUE, border = FALSE, heatscale = c(low = "steelblue", mid =
  "white", high = "tomato"), heatMidpoint = 0, method = "none",
  scaleMax = 2, scaleMin = -2, relative_expr = TRUE, ...)
}
\description{
Plots a pseudotime-ordered, row-centered heatmap
}

