% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot the heatmap with the difference between fitting for branch genes
TO DO: add the confidence interval, add p-val, fix error when we cannot generate fitting for the data}
\usage{
plot_heatmap(abs_branch_gene, cds = absolute_cds, branchA = which(State ==
  2), branchB = which(State == 3), cores = detectCores(),
  trend_formula = "~sm.ns(Pseudotime, df = 3)", fc_limit = 3,
  relative_expr = FALSE, stretch = TRUE, label_by_short_name = TRUE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{relative_expr}{Whether to transform expression into relative values}

\item{rowgenes}{Gene ids or short names to be arrayed on the vertical axis.}

\item{colgenes}{Gene ids or short names to be arrayed on the horizontal axis}

\item{min_expr}{The minimum level of expression to show in the plot}
}
\value{
a ggplot2 plot object
}
\description{
Plot the heatmap with the difference between fitting for branch genes
TO DO: add the confidence interval, add p-val, fix error when we cannot generate fitting for the data
}

