% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ILRs_heatmap}
\alias{plot_ILRs_heatmap}
\title{Create a heatmap to demonstrate the lineage divergence hierarchy for branching genes}
\usage{
plot_ILRs_heatmap(cds, ILRs_df, ABC_df, branching_pval_df,
  rownames_type = c("gene_short_name", "ensemble_id"),
  ABC_type = c("positive", "negative", "all"), dist_method = "euclidean",
  hclust_method = "ward", ILRs_limit = 3, cluster_num = 4, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{ILRs_df}{Matrix of Instant Log Ratio for the branching genes calculated using the calILRs function}

\item{ABC_df}{Matrix of Area Between Curves for the branching genes calculated using the calABCs function}

\item{rownames_type}{The type of row names for the ILRs matrix (either 'gene_short_name' or 'ensemble_id')}

\item{ABC_type}{The type of genes based on ABC score to be selected (either 'positive', 'negative' or 'all')}

\item{dist_method}{The method to calculate distance for each gene used in the hirearchical clustering, any one of them: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski".}

\item{hclust_method}{The method to perform hirearchical clustering, any one of them: ward", "single", "complete", "average", "mcquitty", "median", "centroid"}

\item{ILRs_limit}{the minimum Instant Log Ratio used to make the heatmap plot}

\item{cluster_num}{The minimum level of expression to show in the plot}
}
\value{
a heatmap plot from based on the pheatmap package
}
\description{
Create a heatmap to demonstrate the lineage divergence hierarchy for branching genes
}

