% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_local_spatial_statistics}
\alias{plot_local_spatial_statistics}
\title{plot the local moran's I or the local g for the cells}
\usage{
plot_local_spatial_statistics(cds, markers, method = "local_G", log = TRUE,
  tabulate = TRUE, zero.policy = TRUE, return_all = FALSE, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{method}{a character string specifying the method (the default 'local_G' or 'local_moran') for detecting local spatial statistics along the principal graph embedded in the low dimensional space.}

\item{log}{a logic value to determine whether or not we need to log the gene expression}

\item{tabulate}{A logical argument that when set to TRUE will create column named 'quadrant' in the dataframe supplied to the plot
where the values are either 'high-high', 'high-low', 'low-high', 'low-low' or 'insignificant. The values on the left and right of the hyphen are determined
 by the mean of the value from value column of the dataframe and the mean-moran value respectively.}

\item{zero.policy}{A logic flag that determines the assignment of lagged values of zones without neighbors. When TRUE, zero is assigned. When FALSE NA is assigned. Default value is TRUE.}

\item{return_all}{A logical argument to determine whether or not the dataframe of the local G or Moran's I should be returned}

\item{...}{Additional arguments to pass to scale_color_viridis}
}
\description{
plot the local moran's I or the local g for the cells
}
