% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_markers_cluster}
\alias{plot_markers_cluster}
\title{Heatmap for markers across cell clusters}
\usage{
plot_markers_cluster(cds, markers, sample_cell_num = NULL,
  group_by = "Cluster", show_rownames = TRUE, gene_name_size = 8,
  scale_max = 3, scale_min = -3, minimal_cluster_fraction = NULL,
  return_heatmap = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{markers}{Markers used to create the heatmap. Which is normally ordered for each cluster based on specificity (using dplyr, for example).}

\item{sample_cell_num}{Number of cells to sample. As the single cell RNA-seq datasets get bigger, we cannot visualize all cells. 
This argument helps the users to visualize large scale dataset with an uniform downsampling scheme for cells. By default, if the cell number is larger 
than 50000, we downsample 1000 cells.}

\item{group_by}{The name of the pData column corresponding to the markers.}

\item{show_rownames}{A logic flag to determine whether or not we should show the gene short name}

\item{gene_name_size}{The font size of the gene name}

\item{scale_max}{The maximum value (in standard deviations) to show in the heatmap. Values larger than this are set to the max.}

\item{scale_min}{The minimum value (in standard deviations) to show in the heatmap. Values smaller than this are set to the min.}

\item{minimal_cluster_fraction}{The minimum fraction of cell number for a cluster to the largest cluster to maintain in the heatmap.}

\item{return_heatmap}{A logic flag to determine whether or not we should also return the data frame used to create the heatmap.}

\item{verbose}{Whether or not to show detailed running information}

\item{...}{additional arguments passed into the function. Not used.}
}
\value{
a ggplot2 plot object or a list with the plot object and a "cellDataset" storing the average expression for each cluster of genes across cells
}
\description{
Heatmap for markers across cell clusters
}
