% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_moran}
\alias{plot_moran}
\title{Adapt Moran's I plot}
\usage{
plot_moran(cds, markers, log = TRUE, color_by = "Cluster", quiet = NULL,
  labels = FALSE)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{log}{a logic value to determine whether or not we need to log the gene expression}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{quiet}{default NULL, use !verbose global option value; if TRUE, output of summary of influence object suppressed}

\item{labels}{character labels for points with high influence measures, if set to FALSE, no labels are plotted for points with large influence}
}
\description{
Adapt Moran's I plot
}
