% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_multiple_branches_pseudotime}
\alias{plot_multiple_branches_pseudotime}
\title{Create a kinetic curves to demonstrate the bifurcation of gene expression along multiple branches}
\usage{
plot_multiple_branches_pseudotime(cds, branches, branches_name = NULL,
  min_expr = NULL, cell_size = 0.75, norm_method = c("vstExprs", "log"),
  nrow = NULL, ncol = 1, panel_order = NULL, color_by = "Branch",
  trend_formula = "~sm.ns(Pseudotime, df=3)", label_by_short_name = TRUE,
  TPM = FALSE, cores = 1)
}
\arguments{
\item{cds}{CellDataSet for the experiment (normally only the branching genes detected with branchTest)}

\item{branches}{The ID of the branch point to visualize. Can only be used when reduceDimension is called with method = "DDRTree".}

\item{branches_name}{The two states to compare in the heatmap. Mutually exclusive with branch_point.}

\item{min_expr}{Whether to cluster the rows of the heatmap.}

\item{cell_size}{The method used by pheatmap to perform hirearchical clustering of the rows.}

\item{norm_method}{Number of clusters for the heatmap of branch genes}

\item{nrow}{The color scheme for drawing the heatmap.}

\item{ncol}{Additional annotations to show for each row in the heatmap. Must be a dataframe with one row for each row in the fData table of cds_subset, with matching IDs.}

\item{panel_order}{Additional annotations to show for each column in the heatmap. Must be a dataframe with one row for each cell in the pData table of cds_subset, with matching IDs.}

\item{color_by}{Whether to show the names for each row in the table.}

\item{trend_formula}{A formula string specifying the model used in fitting the spline curve for each gene/feature.}

\item{label_by_short_name}{A logic flag to decide whether or not the gene short name will be used.}

\item{TPM}{Whether to return the pheatmap object to the user.}

\item{cores}{Number of cores to use when smoothing the expression curves shown in the heatmap.}
}
\value{
A list of heatmap_matrix (expression matrix for the branch committment), ph (pheatmap heatmap object),
annotation_row (annotation data.frame for the row), annotation_col (annotation data.frame for the column).
}
\description{
Create a kinetic curves to demonstrate the bifurcation of gene expression along multiple branches
}
