% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ordering_genes}
\alias{plot_ordering_genes}
\title{Plots genes by mean vs. dispersion, highlighting those selected for ordering}
\usage{
plot_ordering_genes(cds)
}
\description{
Plots genes by mean vs. dispersion, highlighting those selected for ordering
}

