% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_PAGA_graph}
\alias{plot_PAGA_graph}
\title{Function used to plot the PAGA graph}
\usage{
plot_PAGA_graph(cds, partition_names = "Cluster", color_by = "Cluster",
  layout = layout.drl, max_vertex_size = 10, max_edge_width = 2,
  vertex_label_cex = 1)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{partition_names}{Which partiton groups (column in the pData) should be used to calculate the connectivity between partitions}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{layout}{A layout function from igraph to layout the PAGA graph}

\item{max_vertex_size}{The maximum value for the vertex size}

\item{max_edge_width}{The maximum value for the edge width}

\item{vertex_label_cex}{The size of vertex label}
}
\description{
Function used to plot the PAGA graph
}
