% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_pseudotime_heatmap}
\alias{plot_pseudotime_heatmap}
\title{Plots a pseudotime-ordered, row-centered heatmap}
\usage{
plot_pseudotime_heatmap(cds_subset,
  trend_formula = "~sm.ns(Pseudotime, df=3)", rescaling = "row",
  clustering = "row", hclust_method = "ward.D2", num_clusters = 6,
  show_rownames = F, logMode = TRUE, pseudocount = 0, use_vst = TRUE,
  border = FALSE, hmcols = NULL, scaleMax = 2, scaleMin = -2,
  return_heatmap = FALSE, relative_expr = TRUE, cores = 1, ...)
}
\arguments{
\item{cds_subset}{CellDataSet for the experiment (normally only the branching genes detected with branchTest)}

\item{hclust_method}{The method to perform hirearchical clustering, any one of them: ward", "single", "complete", "average", "mcquitty", "median", "centroid"}

\item{num_clusters}{Number of clusters for the heatmap of branch genes}

\item{hmcols}{The color scheme for drawing the heatmap}

\item{cores}{Number of cores to run this function}

\item{lineage_labels}{The label for the lineages (first (second) lineage corresponding to state 2 (3))}

\item{norm_method}{Either vstExprs or log, determine which normalization approach will be performed for the fitted gene expression values}

\item{dist_method}{The method to calculate distance for each gene used in the hirearchical clustering, any one of them: "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". (This option is not in function for now)}

\item{heatmap_height}{Height of the saved heatmap}

\item{heatmap_width}{Width of the saved heatmap}

\item{Cell_type_color}{The color for the progenitors and two lineages}

\item{return_all}{}
}
\value{
A list of heatmap_matrix (expression matrix for the lineage committment), ph (pheatmap heatmap object),
annotation_row (annotation data.frame for the row), annotation_col (annotation data.frame for the column). 
Note that, in order to draw the heatmap generate by this function you need to use grid.newpage(); grid.draw(res$ph$gt) (assuming "res" is the variable name for the result of this function)
}
\description{
Plots a pseudotime-ordered, row-centered heatmap
}

