% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ridge}
\alias{plot_ridge}
\title{Plots the percentage of any categories in another group (either categorical or sequential) of cells}
\usage{
plot_ridge(cds, x = "Pseudotime", y = "Cluster", markers = NULL)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{x}{the cell attribute (e.g. the column of pData(cds)) to group cells and use to identify the percentage in}

\item{y}{the cell attribute (e.g. the column of pData(cds)) to group cells and use to identify the percentage to}

\item{markers}{gene of interests used to visualize the gene expression level}
}
\value{
a ggplot2 plot object
}
\description{
Plots the percentage of any categories in another group (either categorical or sequential) of cells
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
HSMM <- reduceDimension(HSMM, reduction_method = 'UMAP')
HSMM <- clusterCells(HSMM)
plot_ridge(HSMM)
}
}
