% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_spanning_tree}
\alias{plot_spanning_tree}
\title{Plots the minimum spanning tree on cells.}
\usage{
plot_spanning_tree(cds, x = 1, y = 2, color_by = "State",
  show_tree = TRUE, show_backbone = TRUE, backbone_color = "black",
  markers = NULL, show_cell_names = FALSE, cell_name_size = 1)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{x}{the column of reducedDimS(cds) to plot on the horizontal axis}

\item{y}{the column of reducedDimS(cds) to plot on the vertical axis}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{show_tree}{whether to show the links between cells connected in the minimum spanning tree}

\item{show_backbone}{whether to show the diameter path of the MST used to order the cells}

\item{backbone_color}{the color used to render the backbone.}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{show_cell_names}{draw the name of each cell in the plot}

\item{cell_name_size}{the size of cell name labels}
}
\value{
a ggplot2 plot object
}
\description{
Plots the minimum spanning tree on cells.
}
\examples{
\dontrun{
data(HSMM)
plot_spanning_tree(HSMM)
plot_spanning_tree(HSMM, color_by="Pseudotime", show_backbone=FALSE)
plot_spanning_tree(HSMM, markers="MYH3")
}
}

