% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order_cells.R
\name{pq_helper}
\alias{pq_helper}
\title{Recursively builds and returns a PQ tree for the MST}
\usage{
pq_helper(mst, use_weights = TRUE, root_node = NULL)
}
\description{
Recursively builds and returns a PQ tree for the MST
}

