% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{procrustes}
\alias{procrustes}
\title{The following functioin is taken from vegan package for performing procrustes analysis}
\usage{
procrustes(X, Y, scale = TRUE, symmetric = FALSE)
}
\arguments{
\item{X}{Target matrix}

\item{Y}{Matrix to be rotated.}

\item{scale}{Allow scaling of axes of Y.}

\item{symmetric}{Use symmetric Procrustes statistic (the rotation will still be non-symmetric).}
}
\description{
Function procrustes rotates a configuration to maximum similarity with another configuration. Function protest tests the non-randomness (significance) between two configurations.
}
