% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order_cells.R
\name{project_point_to_line_segment}
\alias{project_point_to_line_segment}
\title{Project point to line segment}
\usage{
project_point_to_line_segment(p, df)
}
\description{
Project point to line segment
}

