% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{project_to_representatives}
\alias{project_to_representatives}
\title{apply force-directed layout on the kNN graph built from the downsampled representive cells and then project other 
non-representative cells to the low dimensional space with five nearest (on original UMAP space) representive cells' coordinates}
\usage{
project_to_representatives(data, data_ori, landmark_id, louvain_res, pd,
  method = "drl", start.temp = NULL, verbose, ...)
}
\arguments{
\item{data}{(dowmsampled) data to perform Louvain clustering and kNN graph construction}

\item{data_ori}{All data points without downsampling. This data space will be used to find the nearest five points for projecting non-landmark points}

\item{landmark_id}{The index of cells}

\item{louvain_res}{The result of louvain clustering clustering from the original space}

\item{pd}{the data frame of phenotype information}

\item{method}{The force directed layout function to use}

\item{start.temp}{argument passed into layout_with_fr function}

\item{verbose}{Wheter to print all running details}

\item{...}{additional arguments passed to functions (louvain_clustering) called by this function.}
}
\description{
apply force-directed layout on the kNN graph built from the downsampled representive cells and then project other 
non-representative cells to the low dimensional space with five nearest (on original UMAP space) representive cells' coordinates
}
