% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{projectPCA}
\alias{projectPCA}
\title{project a CellDataSet object into a lower dimensional PCA space}
\usage{
projectPCA(cds, num_dim = 50, norm_method = c("log", "vstExprs", "none"),
  residualModelFormulaStr = NULL, pseudo_expr = 1, relative_expr = TRUE,
  auto_param_selection = TRUE, verbose = FALSE, scaling = TRUE, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{num_dim}{the dimensionality of the reduced space}

\item{norm_method}{Determines how to transform expression values prior to reducing dimensionality}

\item{residualModelFormulaStr}{A model formula specifying the effects to subtract from the data before clustering.}

\item{pseudo_expr}{amount to increase expression values before dimensionality reduction}

\item{relative_expr}{When this argument is set to TRUE (default), we intend to convert the expression into a relative expression.}

\item{verbose}{Whether to emit verbose output during dimensionality reduction}

\item{scaling}{When this argument is set to TRUE (default), it will scale each gene before running trajectory reconstruction.}

\item{...}{additional arguments to pass to the dimensionality reduction function}
}
\value{
an updated CellDataSet object
}
\description{
For most analysis (including trajectory inference, clustering) in Monocle 3, it requires us to to start from a 
low dimensional PCA space. projectPCA will be used to first project a CellDataSet object into a lower dimensional PCA space 
before we apply clustering with community detection algorithm or other non-linear dimension reduction method, for example 
UMAP, tSNE, DDRTree, SSE, L1-graph, SGL-tree, etc.  
While tSNE is especially suitable for visualizing clustering result, comparing to UMAP, the global distance in tSNE space is 
not meaningful. UMAP can either be used for visualizing clustering result or as a general non-linear dimension reduction method. 
It can be used in conjunction with SSE to obtain smooth skeleton representation of the data. DDRTree and L1-graph are two complementary 
trajectory inference method where the first one is very great at learning a tree structure but the later is general and can 
learn any arbitrary graph structure. Both methods can be applied to the UMAP space or the smoothed SSE space.
}
