% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\name{reducedDimA<-}
\alias{reducedDimA<-}
\title{Sets the weights transform the cells' coordinates in the reduced dimension space back to the full (whitened) space.}
\usage{
reducedDimA(cds) <- value
}
\arguments{
\item{cds}{A CellDataSet object.}

\item{value}{A whitened expression data matrix}
}
\value{
An updated CellDataSet object
}
\description{
Sets the weights transform the cells' coordinates in the reduced dimension space back to the full (whitened) space.
}
\examples{
\dontrun{
cds <- reducedDimA(A)
}
}

