% Generated by roxygen2 (4.0.0): do not edit by hand
\name{reducedDimS<-}
\alias{reducedDimS<-}
\title{Sets the coordinates of each cell in the reduced-dimensionality space.  Not intended to be called directly.}
\usage{
reducedDimS(cds) <- value
}
\arguments{
\item{cds}{A CellDataSet object.}

\item{value}{A matrix of coordinates specifying each cell's position in the reduced-dimensionality space.}
}
\value{
An update CellDataSet object
}
\description{
Sets the coordinates of each cell in the reduced-dimensionality space.  Not intended to be called directly.
}

