% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reducedDimS}
\alias{reducedDimS}
\title{Retrieves the coordinates of each cell in the reduced-dimensionality space generated by calls to
reduceDimension.}
\usage{
reducedDimS(cds)
}
\arguments{
\item{cds}{A CellDataSet object.}
}
\value{
A matrix, where rows are cell coordinates and columns correspond to dimensions of the
reduced space.
}
\description{
Reducing the dimensionality of the expression data is a core step in the Monocle
workflow. After you call reduceDimension(), this function will return the new
coordinates of your cells in the reduced space.
}
\examples{
\dontrun{
data(HSMM)
S <- reducedDimS(HSMM)
}
}

