% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{reduceDimension}
\alias{reduceDimension}
\title{Computes a projection of a CellDataSet object into a lower dimensional space}
\usage{
reduceDimension(cds, max_components = 2, reduction_method = c("DDRTree",
  "ICA"), pseudo_expr = NULL, residualModelFormulaStr = NULL,
  use_vst = NULL, verbose = FALSE, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{max_components}{the dimensionality of the reduced space}

\item{reduction_method}{A character string specifying the algorithm to use for dimensionality reduction.}

\item{pseudo_expr}{amount to increase expression values before dimensionality reduction}

\item{residualModelFormulaStr}{A model formula specifying the effects to subtract from the data before clustering.}

\item{use_vst}{Whether to variance-stabilize the expression data before reduction.
Defaults to TRUE if the expressionFamily of cds is negative binomial, FALSE otherwise. 
Only works for negative binomial CellDataSets objects.}

\item{verbose}{Whether to emit verbose output during dimensionality reduction}

\item{...}{additional arguments to pass to the dimensionality reduction function}
}
\value{
an updated CellDataSet object
}
\description{
Computes a projection of a CellDataSet object into a lower dimensional space
}

