% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{residualMatrix}
\alias{residualMatrix}
\title{Response values}
\usage{
residualMatrix(models, residual_type = "response", cores = 1)
}
\arguments{
\item{models}{a list of models, e.g. as returned by fitModels()}

\item{residual_type}{the response desired, as accepted by VGAM's predict function}

\item{cores}{number of cores used for calculation}
}
\value{
a matrix where each row is a vector of response values for a particular feature's model, and columns are cells.
}
\description{
Generates a matrix of response values for a set of fitted models
}
