% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{reverseEnbedingCDS}
\alias{reverseEnbedingCDS}
\title{Reverse enbedding latent graph coordinates back to the high dimension}
\usage{
reverseEnbedingCDS(cds)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}
}
\value{
a new cds containing only the genes used in reducing dimension. Expression values are reverse embedded.
}
\description{
Reverse enbedding latent graph coordinates back to the high dimension
}

