% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{run_dpt}
\alias{run_dpt}
\title{Run dpt for dimension reduction}
\usage{
run_dpt(data, branching = T, norm_method = "log", root = NULL,
  verbose = F)
}
\arguments{
\item{branching}{A logic argument to determine whether or not branch detection should turn on in DPT function}

\item{norm_method}{A character argument to determine which normalization method used for preprocessing the data}

\item{root}{Index to the root cell, if no root cell set, a random cell will be picked up}

\item{verbose}{A logic argument to determine whether or not to print running message}

\item{cds}{the CellDataSet upon which to perform this operation}
}
\value{
a list
}
\description{
This function perform dimension reduction with dpt
}

