% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{selectFeatureByDP}
\alias{selectFeatureByDP}
\title{Select features for constructing the developmental trajectory}
\usage{
selectFeatureByDP(cds, num_cells_expressed = 5, num_dim = 5,
  rho_threshold = NULL, delta_threshold = NULL, qval_threshold = 0.01,
  verbose = F)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{num_cells_expressed}{A logic flag to determine whether or not you want to skip the calculation of rho / sigma}

\item{num_dim}{Number of clusters. The algorithm use 0.5 of the rho as the threshold of rho and the delta 
corresponding to the number_clusters sample with the highest delta as the density peaks and for assigning clusters}

\item{rho_threshold}{The threshold of local density (rho) used to select the density peaks}

\item{delta_threshold}{The threshold of local distance (delta) used to select the density peaks}

\item{qval_threshold}{A logic flag passed to densityClust function in desnityClust package to determine whether or not Gaussian kernel will be used for calculating the local density}

\item{verbose}{Verbose parameter for DDRTree}
}
\value{
an list contains a updated CellDataSet object after clustering and tree construction as well as a vector including the selected top significant differentially expressed genes across clusters of cells
}
\description{
Select features for constructing the developmental trajectory
}
\references{
Rodriguez, A., & Laio, A. (2014). Clustering by fast search and find of density peaks. Science, 344(6191), 1492-1496. doi:10.1126/science.1242072
}

