% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{selectGenesInExpressionRange}
\alias{selectGenesInExpressionRange}
\title{Filter genes outside of a given range of expression}
\usage{
selectGenesInExpressionRange(cds, min_expression_threshold = -Inf,
  max_expression_threshold = Inf, detectionLimit = -Inf,
  stat_fun = median, relative_expr = TRUE)
}
\description{
Filter genes outside of a given range of expression
}

