% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{selectHighDispersionGenes}
\alias{selectHighDispersionGenes}
\title{Filter genes outside of a given range of expression}
\usage{
selectHighDispersionGenes(cds, dispersionThresh = 0.1, relative_expr = TRUE)
}
\description{
Filter genes outside of a given range of expression
}

