% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{selectTrajectoryRoots}
\alias{selectTrajectoryRoots}
\title{Select the roots of the principal graph}
\usage{
selectTrajectoryRoots(cds, x = 1, y = 2, num_roots = NULL, pch = 19,
  ...)
}
\arguments{
\item{cds}{CellDataSet where roots will be selected from}

\item{pch}{}
}
\description{
Select the roots of the principal graph
}
