% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{shannon.entropy}
\alias{shannon.entropy}
\title{Calculate the Shannon entropy based on the probability vector}
\usage{
shannon.entropy(p)
}
\description{
Calculate the Shannon entropy based on the probability vector
}

