% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{SubSet_cds}
\alias{SubSet_cds}
\title{Subset a cds which only includes cells provided with the argument cells}
\usage{
SubSet_cds(cds, cells)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}

\item{cells}{a vector contains all the cells you want to subset}
}
\value{
a new cds containing only the cells from the cells argument
}
\description{
Subset a cds which only includes cells provided with the argument cells
}

