% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{subsetCDS}
\alias{subsetCDS}
\title{Subset a cds which only includes cells provided with the argument cells}
\usage{
subsetCDS(cds, cells, principal_nodes = NULL)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}

\item{cells}{a vector contains all the cells you want to subset}
}
\value{
a new cds containing only the cells from the cells argument
}
\description{
Subset a cds which only includes cells provided with the argument cells
}
\examples{
\dontrun{
lung <- load_lung()
tmp <- subset_cds(lung, cells = row.names(subset(pData(lung), State == 1)))
plot_cell_trajectory(tmp)
}
}
