% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tenx_to_cds}
\alias{tenx_to_cds}
\title{This function reads in a list of 10X pipeline output directories and output a Monocle cell dataset for downstream analysis}
\usage{
tenx_to_cds(pipeline_dirs, genome = "hg19", include_analysis = F,
  lowerDetectionLimit = 1)
}
\arguments{
\item{pipeline_dirs}{Directory name or list of directory names of the top level 10X output directory for an experiment(s)}

\item{genome}{String with genome name specified for 10X run (such as hg19)}

\item{include_analysis}{A boolean that signifies whether or not an analysis path can be found in the 10X output and if you wish to include it}

\item{lowerDetectionLimit}{A number that signifies the minimum expression level required in order for a gene to be truly expressed in a cell}
}
\value{
A cellDataSet object containing data from all experiments.
}
\description{
Takes a list of 10X pipeline output directories and generates a cellDataSet containing all cells in these experiments. 
This function is originally from Andrew Hill.
}
\examples{
\dontrun{
cds <- tenx_to_cds(c("/net/trapnell/vol1/home/xqiu/aggregated_samples_1", 
                        "/net/trapnell/vol1/home/xqiu/aggregated_samples_1", 'hg19'))
}
}
