% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{traverseGraphCDS}
\alias{traverseGraphCDS}
\title{Make a cds by traversing from one cell to another cell}
\usage{
traverseGraphCDS(cds, interactive = TRUE, starting_cell = NULL,
  end_cells = NULL, ...)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}

\item{interactive}{whether to run this function in an interactive mode. If running with an interactive mode, arguments starting_cell and end_cells will be ignored.}

\item{starting_cell}{the initial vertex for traversing on the graph}

\item{end_cells}{the terminal vertex for traversing on the graph}
}
\value{
a new cds containing only the cells traversed from the intial cell to the end cell
}
\description{
Make a cds by traversing from one cell to another cell
}
\keyword{internal}
