% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{traverseTreeCDS}
\alias{traverseTreeCDS}
\title{Make a cds by traversing from one cell to another cell}
\usage{
traverseTreeCDS(cds, starting_cell, end_cells)
}
\arguments{
\item{cds}{a cell dataset after trajectory reconstruction}

\item{starting_cell}{the initial vertex for traversing on the graph}

\item{end_cells}{the terminal vertex for traversing on the graph}
}
\value{
a new cds containing only the cells traversed from the intial cell to the end cell
}
\description{
Make a cds by traversing from one cell to another cell
}
