% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellDataSet.R
\name{updateCDS}
\alias{updateCDS}
\title{Functions to update a old CDS to the current version}
\usage{
updateCDS(cds)
}
\arguments{
\item{cds}{CellDataSet you'd like to update}
}
\description{
Functions to update a old CDS to the current version
}
