% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{vstExprs}
\alias{vstExprs}
\title{Return a variance-stabilized matrix of expression values}
\usage{
vstExprs(cds, dispModelName = "blind", expr_matrix = NULL, round_vals = TRUE)
}
\arguments{
\item{cds}{A CellDataSet to use for variance stabilization.}

\item{dispModelName}{The name of the dispersion function to use for VST.}

\item{expr_matrix}{An matrix of values to transform. Must be normalized (e.g. by size factors) already. This function doesn't do this for you.}

\item{round_vals}{Whether to round expression values to the nearest integer before applying the transformation.}
}
\description{
This function was taken from the DESeq package (Anders and Huber) and modified
to suit Monocle's needs. It accpets a either a CellDataSet or the expression values of one
and returns a variance-stabilized matrix based off of them.
}
