% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM.R
\name{BEAM}
\alias{BEAM}
\title{Peform the beam analysis test}
\usage{
BEAM(cds, fullModelFormulaStr = "~sm.ns(Pseudotime, df = 3)*Lineage",
  reducedModelFormulaStr = "~sm.ns(Pseudotime, df = 3)",
  lineage_states = NULL, branch_point = 1, relative_expr = TRUE,
  stretch = TRUE, q_thrsld = 0.05, lineage_labels = NULL,
  verbose = FALSE, cores = 1, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{fullModelFormulaStr}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{reducedModelFormulaStr}{a formula string specifying the reduced model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{lineage_states}{ids for the immediate branch lineage which obtained from lineage construction based on MST}

\item{branch_point}{The ID of the branch point to analyze. Can only be used when reduceDimension is called with method = "DDRTree".}

\item{relative_expr}{a logic flag to determine whether or not the relative gene expression should be used}

\item{stretch}{a logic flag to determine whether or not each lineage should be stretched}

\item{q_thrsld}{The threshold for the qval}

\item{lineage_labels}{the name for each lineage, for example, "AT1" or "AT2"}

\item{verbose}{Whether to generate verbose output}

\item{cores}{the number of cores to be used while testing each gene for differential expression}

\item{...}{additional arguments to be passed to differentialGeneTest}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models, as well as the time point where the gene starts to bifurcate between two branches
}
\description{
Perform BEAM analysis
}

