% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{branchTest}
\alias{branchTest}
\title{find the branch genes}
\usage{
branchTest(cds, fullModelFormulaStr = "~Lineage * ns(Pseudotime, df = 3)",
  reducedModelFormulaStr = "~ns(Pseudotime, df = 3)", lineage_states = c(2,
  3), stretch = FALSE, cores = 1)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{fullModelFormulaStr}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{reducedModelFormulaStr}{a formula string specifying the reduced model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{lineage_states}{ids for the immediate branch lineage which obtained from lineage construction based on MST}

\item{cores}{the number of cores to be used while testing each gene for differential expression}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
find the branch genes
}

