% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{branchTest}
\alias{branchTest}
\title{Peform the branching test}
\usage{
branchTest(cds, fullModelFormulaStr = "~sm.ns(Pseudotime, df = 3)*Lineage",
  reducedModelFormulaStr = "~sm.ns(Pseudotime, df = 3)",
  lineage_states = c(2, 3), relative_expr = TRUE, stretch = TRUE,
  pseudocount = 0, cores = 1, weighted = TRUE, lineage_labels = NULL,
  ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{fullModelFormulaStr}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{reducedModelFormulaStr}{a formula string specifying the reduced model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{lineage_states}{ids for the immediate branch lineage which obtained from lineage construction based on MST}

\item{relative_expr}{a logic flag to determine whether or not the relative gene expression should be used}

\item{stretch}{a logic flag to determine whether or not each lineage should be stretched}

\item{pseudocount}{pseudo count added before fitting the spline curves}

\item{cores}{the number of cores to be used while testing each gene for differential expression}

\item{weighted}{A logic flag to determine whether or not we should use the navie logLikelihood weight scheme for the duplicated progenitor cells}

\item{lineage_labelsthe}{name for each lineage, for example, AT1 or AT2}

\item{gene_names}{gene names used to make the bifurcation plots for two genes.}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
This function is used to perform the branching test to ask the question about whether or not the genes under tests are significant
lineage dependent genes. If stretch equal to TRUE, each lineage is firstly stretched into maturation level 0-100 and the progenitor
cells are duplicated and assigned to each lineage. This test can be used to detect lineage dependent genes.
}

