% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM.R
\name{buildLineageBranchCellDataSet}
\alias{buildLineageBranchCellDataSet}
\title{Build a CellDataSet with appropriate duplication along two lineages}
\usage{
buildLineageBranchCellDataSet(cds, progenitor_method = c("sequential_split",
  "random_split", "duplicate"), lineage_states = NULL, branch_point = 1,
  lineage_labels = NULL, stretch = TRUE)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{progenitor_method}{The method to treat the progenitor cells, there are three options: sequential_split, random_split, duplicate. 
They represent that the progenitors can be either splitted one by one (sequential_split) or randomly (random_split) or duplicate the 
progenitors (duplicate). For random_split and sequential_split, the cells will be evenly split (when number of progenitors are odd, 
there will be one more cell in one lineage)}

\item{lineage_states}{The states for two branching lineages}

\item{branch_point}{The ID of the branch point to analyze. Can only be used when reduceDimension is called with method = "DDRTree".}

\item{lineage_labels}{The names for each branching lineage}

\item{stretch}{A logic flag to determine whether or not the pseudotime trajectory for each lineage should be stretched to the same range or not}
}
\value{
a CellDataSet with the duplicated cells and stretched lineages
}
\description{
Build a CellDataSet with appropriate duplication along two lineages
}

