% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM.R
\name{buildLineageBranchCellDataSet}
\alias{buildLineageBranchCellDataSet}
\title{Build a CellDataSet with appropriate duplication along two lineages}
\usage{
buildLineageBranchCellDataSet(cds, lineage_states = c(2, 3),
  lineage_labels = NULL, stretch = FALSE, weighted = TRUE, ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{lineage_states}{The states for two branching lineages}

\item{lineage_labels}{The names for each branching lineage}

\item{stretch}{A logic flag to determine whether or not the pseudotime trajectory for each lineage should be stretched to the same range or not}

\item{weighted}{A logic flag to determine whether or not we should use the navie logLikelihood weight scheme for the duplicated progenitor cells}
}
\value{
a CellDataSet with the duplicated cells and stretched lineages
}
\description{
Build a CellDataSet with appropriate duplication along two lineages
}

