% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{cal_ncenter}
\alias{cal_ncenter}
\title{Function to decide a good number of centers for running DDRTree on big datasets}
\usage{
cds includeDescrip will place a better description when warnings no longer appear
}
\arguments{
\item{ncells}{includeDescrip}

\item{ncells_limit}{includeDescrip}

\item{cds}{a cell dataset after trajectory reconstruction}
}
\value{
a new cds containing only the genes used in reducing dimension. Expression values are reverse embedded.
}
\description{
Function to decide a good number of centers for running DDRTree on big datasets
}
