% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{calABCs}
\alias{calABCs}
\title{calculate the area between TWO fitted lineage trajectories}
\usage{
calABCs(cds, fullModelFormulaStr = "~sm.ns(Pseudotime, df = 3)*Lineage",
  reducedModelFormulaStr = "~sm.ns(Pseudotime, df = 3)", method = "fitting",
  ABC_method = "integral", points_num = 1000, fc_limit = 3,
  branchTest = FALSE, lineage_states = c(2, 3), relative_expr = TRUE,
  stretch = TRUE, pseudocount = 0, cores = 1, weighted = TRUE,
  min_expr = 0.5, integer_expression = FALSE, num = 5000,
  lineage_labels = NULL, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{fullModelFormulaStr}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{reducedModelFormulaStr}{a formula string specifying the reduced model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{branchTest}{a logic flag to determine whether or not to perform the branchTest inside the function. Because of the long computations, we recommend to first perform the branchTest and then calculate the ABCs for the genes of interests. Otherwise the ABCs will be appended to the last column of branchTest results.}

\item{lineage_states}{ids for the immediate branch lineage which obtained from lineage construction based on MST (only two lineages are allowed for this function)}

\item{cores}{the number of cores to be used while testing each gene for differential expression}

\item{min_expr}{the lower limit for the expressed gene}

\item{integer_expression}{the logic flag to determine whether or not the integer numbers are used for calculating the ABCs. Default is False.}

\item{num}{number of points on the fitted lineage trajectories used for calculating the ABCs. Default is 5000.}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
calculate the area between TWO fitted lineage trajectories
}

