% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{calculateMarkerSpecificity}
\alias{calculateMarkerSpecificity}
\title{Calculate each gene's specificity for each cell type}
\usage{
calculateMarkerSpecificity(cds, cth, ...)
}
\description{
Calculate each gene's specificity for each cell type
}

