% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calibrate_mc}
\alias{calibrate_mc}
\title{Make an educated guess on the spike-based slope regression parameters}
\usage{
calibrate_mc(total_mRNA, capture_rate, ladder, total_ladder_transcripts,
  trials = 100)
}
\description{
Make an educated guess on the spike-based slope regression parameters
}

