% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{calibrate_mode}
\alias{calibrate_mode}
\title{Function used to calibrate the mode, m and c as well as the total mRNA}
\usage{
calibrate_mode(ind, tpm_distribution, ladder, total_ladder_transcripts,
  total_mRNA, capture_rate, reads, trials = 100)
}
\arguments{
\item{ind}{a number (ind is short for specified "index")}

\item{tpm_distribution}{test}

\item{ladder}{test}

\item{total_ladder_transcripts}{test}

\item{total_mRNA}{a number}

\item{capture_rate}{a number}

\item{reads}{a number}

\item{trials}{a number}
}
\description{
Function used to calibrate the mode, m and c as well as the total mRNA
}

