% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEAM.R
\name{calILRs}
\alias{calILRs}
\title{Calculate the Instant Log Ratio between two branching lineages}
\usage{
calILRs(cds, trajectory_states = c(2, 3), lineage_labels = NULL,
  stretch = T, cores = 1,
  trend_formula = "~sm.ns(Pseudotime, df = 3)*Lineage", ILRs_limit = 3,
  relative_expr = TRUE, weighted = FALSE, label_by_short_name = TRUE,
  useVST = FALSE, round_exprs = FALSE, pseudocount = 0,
  output_type = "all", file = NULL, return_all = F, verbose = FALSE,
  ...)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{cores}{Number of cores when fitting the spline curves}

\item{trend_formula}{the model formula to be used for fitting the expression trend over pseudotime}

\item{ILRs_limit}{the minimum Instant Log Ratio used to make the heatmap plot}

\item{relative_expr}{A logic flag to determine whether or not the relative expressed should be used when we fitting the spline curves}

\item{weighted}{A logic flag to determine whether or not we should use the navie logLikelihood weight scheme for the duplicated progenitor cells}

\item{label_by_short_name}{label the rows of the returned matrix by gene_short_name (TRUE) or feature id (FALSE)}

\item{useVST}{A logic flag to determine whether or not the Variance Stablization Transformation should be used to stablize the gene expression.
When VST is used, the difference between two lineages are used instead of the log-ratio.}

\item{round_exprs}{A logic flag to determine whether or not the expression value should be rounded into integer}

\item{pseudocount}{pseudo count added before fitting the spline curves}

\item{output_type}{A character either of "all" or "after_bifurcation". If "after_bifurcation" is used, only the time points after the bifurcation point will be selected}

\item{file}{the name for storing the data. Since the calculation of the Instant Log Ratio is very time consuming, so by default the result will be stored}

\item{Lineage}{The column in pData used for calculating the ILRs (If not equal to "Lineage", a warning will report)}

\item{lineage_states}{The states for two branching lineages}
}
\value{
a ggplot2 plot object
}
\description{
This function is used to calculate the Instant Log Ratio between two branching lineages which can be used to prepare the heatmap demonstrating the lineage gene expression divergence hirearchy. If "stretch" is specifified, each  
lineage will be firstly stretched into maturation level from 0-100. Since the results when we use "stretching" are always better and 
IRLs for non-stretched spline curves are often mismatched, we may only turn down "non-stretch" functionality in future versions. Then, we fit two separate nature spline curves for each 
individual linages. The log-ratios of the value on each spline curve corresponding to each lineages are calculated, which can be  
used as a measure for the magnitude of divergence between two branching lineages.
}

