% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\docType{methods}
\name{CellDataSet-methods}
\alias{CellDataSet,ANY,ANY-method}
\alias{CellDataSet-methods}
\alias{cellPairwiseDistances}
\alias{cellPairwiseDistances<-}
\alias{estimateDispersions,CellDataSet-method}
\alias{estimateSizeFactors,CellDataSet-method}
\alias{minSpanningTree}
\alias{minSpanningTree<-}
\alias{reducedDimA}
\alias{reducedDimA<-}
\alias{reducedDimK}
\alias{reducedDimK<-}
\alias{reducedDimW<-}
\alias{sizeFactors,CellDataSet-method}
\alias{sizeFactors<-,CellDataSet,numeric-method}
\title{Methods for the CellDataSet class}
\usage{
\S4method{sizeFactors}{CellDataSet}(object)

\S4method{sizeFactors}{CellDataSet,numeric}(object) <- value

\S4method{estimateSizeFactors}{CellDataSet}(object, locfunc = median, ...)

\S4method{estimateDispersions}{CellDataSet}(object,
  modelFormulaStr = "expression~1", relative_expr = TRUE, cores = 1,
  dispModelName = "blind", ...)

reducedDimK(cds) <- value

reducedDimK(cds)

reducedDimW(cds) <- value

reducedDimA(cds)

reducedDimA(cds) <- value

minSpanningTree(cds)

minSpanningTree(cds) <- value

cellPairwiseDistances(cds)

cellPairwiseDistances(cds) <- value
}
\arguments{
\item{value}{a numeric matrix}

\item{cds}{A CellDataSet object.}

\item{cds}{A CellDataSet object.}

\item{cds}{A CellDataSet object.}

\item{value}{A whitened expression data matrix}

\item{cds}{A CellDataSet object.}

\item{cds}{A CellDataSet object.}

\item{value}{A whitened expression data matrix}

\item{cds}{expression data matrix for an experiment}

\item{cds}{A CellDataSet object.}

\item{value}{an igraph object describing the minimum spanning tree.}

\item{cds}{expression data matrix for an experiment}

\item{cds}{A CellDataSet object.}

\item{value}{a square, symmetric matrix containing pairwise distances between cells.}
}
\value{
A matrix, where each row is a set of whitened expression values for a feature and columns are cells.

A matrix, where each row is a set of whitened expression values for a feature and columns are cells.

An updated CellDataSet object

A matrix that when multiplied by a reduced-dimension set of coordinates for the CellDataSet,
recovers a matrix in the full (whitened) space

An updated CellDataSet object

An igraph object representing the CellDataSet's minimum spanning tree.

An updated CellDataSet object

A square, symmetric matrix containing the distances between each cell in the reduced-dimensionality space.

An updated CellDataSet object
}
\description{
Methods for the CellDataSet class

Sets the the whitening matrix during independent component analysis.

Retrieves the the whitening matrix during independent component analysis.

Sets the whitened expression values for each cell prior to independent component analysis. Not intended to be called directly.

Retrieves the weights that transform the cells' coordinates in the reduced dimension space back to the full (whitened) space.

Sets the weights transform the cells' coordinates in the reduced dimension space back to the full (whitened) space.

Retrieves the minimum spanning tree generated by Monocle during cell ordering.

Sets the minimum spanning tree used by Monocle during cell ordering. Not intended to be called directly.

Retrieves a matrix capturing distances between each cell in the reduced-dimensionality space

Sets the matrix containing distances between each pair of cells used by Monocle during cell ordering. Not intended to be called directly.
}
\examples{
data(HSMM)
W <- reducedDimW(HSMM)
data(HSMM)
W <- reducedDimW(HSMM)
data(HSMM)
A <- reducedDimA(HSMM)
data(HSMM)
T <- minSpanningTree(HSMM)
data(HSMM)
D <- cellPairwiseDistances(HSMM)
}

