% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\docType{methods}
\name{CellDataSet-methods}
\alias{CellDataSet,ANY,ANY-method}
\alias{CellDataSet-methods}
\alias{estimateDispersions,CellDataSet-method}
\alias{estimateSizeFactors,CellDataSet-method}
\alias{sizeFactors,CellDataSet-method}
\alias{sizeFactors<-,CellDataSet,numeric-method}
\title{Methods for the CellDataSet class}
\usage{
\S4method{sizeFactors}{CellDataSet}(object)

\S4method{sizeFactors}{CellDataSet,numeric}(object) <- value

\S4method{estimateSizeFactors}{CellDataSet}(object, locfunc = median, ...)

\S4method{estimateDispersions}{CellDataSet}(object, modelFormulaStr = "~ 1",
  relative_expr = TRUE, cores = 1, dispModelName = "blind", ...)
}
\arguments{
\item{relative_expr}{Whether to transform expression into relative values}

\item{A}{function applied to the geometric-mean-scaled expression values to derive the size factor.}
}
\description{
Methods for the CellDataSet class
}

