% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellDataSet.R
\name{cellPairwiseDistances<-}
\alias{cellPairwiseDistances<-}
\title{Sets the matrix containing distances between each pair of cells used by Monocle during cell ordering. Not intended to be called directly.}
\usage{
cellPairwiseDistances(cds) <- value
}
\arguments{
\item{cds}{A CellDataSet object.}

\item{value}{a square, symmetric matrix containing pairwise distances between cells.}
}
\value{
An updated CellDataSet object
}
\description{
Sets the matrix containing distances between each pair of cells used by Monocle during cell ordering. Not intended to be called directly.
}
\examples{
\dontrun{
cds <- cellPairwiseDistances(D)
}
}
