% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{classifyCellClusters}
\alias{classifyCellClusters}
\title{classifyCellClusters}
\usage{
classifyCellClusters(cds, cth, frequency_thresh = 0.8,
  grouping_var = "Cluster")
}
\description{
Impute cell type classifications using clustering results
}
