% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{classifyCellsGlmNet}
\alias{classifyCellsGlmNet}
\title{classifyCellsGlmNet}
\usage{
classifyCellsGlmNet(cds, cth, rank_prob_ratio = 2, min_observations = 8,
  max_training_samples = 10000, cores = 1)
}
\arguments{
\item{cds}{CellDataSet containing cells that will be clustered}

\item{cth}{CellTypeHierarchy that dictates cell types present and requirements of cell to be considered  a certain cell type}

\item{rank_prob_ratio}{The probability ratio of the rank}

\item{min_observations}{Minimun of the observation}

\item{max_training_samples}{Maximum training samples}

\item{cores}{Number of cores computer should use to execute function}
}
\description{
Description of classifyCellsGlmNet function -- To be added
}
