% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{classifyCellsHelperCellGlmNet}
\alias{classifyCellsHelperCellGlmNet}
\title{classifyCellsHelperCellGlmNet}
\usage{
classifyCellsHelperCellGlmNet(cds, cth)
}
\arguments{
\item{cds}{CellDataSet containing cells that will be clustered}

\item{cth}{CellTypeHierarchy that dictates cell types present and requirements of cell to be considered  a certain cell type}
}
\description{
Description of classifyCellsHelperCellGlmNet -- to be added
}
