% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_cells.R
\name{clusterCells}
\alias{clusterCells}
\title{Cluster cells to nearest DDRTree center.}
\usage{
clusterCells(cds, num_clusters = 2, num_reduced_dims = 10,
  residualModelFormulaStr = NULL, ddrtree_gamma = 100)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{num_clusters}{number of desired cell clusters}

\item{num_reduced_dims}{number of dimensions to reduce the dataset to}

\item{ddrtree_gamma}{gamma parameter for DDRTree}
}
\value{
an updated CellDataSet object, in which phenoData contains values for Cluster for each cell
}
\description{
Cluster cells to nearest DDRTree center.
}

