% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterCells}
\alias{clusterCells}
\title{Cluster cells into a specified number of groups.}
\usage{
clusterCells(cds, num_clusters, cell_type_hierarchy = NULL,
  frequency_thresh = 0.1, clustering_genes = NULL, num_reduced_dims = 10,
  residualModelFormulaStr = NULL, ddrtree_gamma = 100, verbose = F)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{num_clusters}{number of desired cell clusters}

\item{num_reduced_dims}{number of dimensions to reduce the dataset to}

\item{residualModelFormulaStr}{A model formula specifying the effects to subtract from the data before clustering.}

\item{ddrtree_gamma}{gamma parameter for DDRTree}

\item{verbose}{Verbose parameter for DDRTree}
}
\value{
an updated CellDataSet object, in which phenoData contains values for Cluster for each cell
}
\description{
Unsupervised clustering of cells is a common step in many single-cell expression
workflows. In an experiment containing a mixture of cell types, each cluster might
correspond to a different cell type. This method takes a CellDataSet as input
along with a requested number of clusters, clusters them with an unsupervised 
algorithm, and then returns the CellDataSet with the cluster assignments stored in
the pData table.
}

