% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{diff_test_helper}
\alias{diff_test_helper}
\title{Helper function for parallel differential expression testing}
\usage{
diff_test_helper(
  x,
  fullModelFormulaStr,
  reducedModelFormulaStr,
  expressionFamily,
  relative_expr,
  weights,
  disp_func = NULL,
  verbose = FALSE
)
}
\arguments{
\item{x}{test}

\item{fullModelFormulaStr}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{reducedModelFormulaStr}{a formula string specifying the reduced model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{expressionFamily}{specifies the VGAM family function used for expression responses}

\item{relative_expr}{Whether to transform expression into relative values}

\item{weights}{test}

\item{disp_func}{test}

\item{verbose}{Whether to show VGAM errors and warnings. Only valid for cores = 1.}
}
\description{
test
}
