% Generated by roxygen2 (4.0.2): do not edit by hand
\name{differentialGeneTest}
\alias{differentialGeneTest}
\title{Tests each gene for differential expression as a function of progress through a biological process, or according to other covariates as specified.}
\usage{
differentialGeneTest(cds,
  fullModelFormulaStr = "expression~sm.ns(Pseudotime, df=3)",
  reducedModelFormulaStr = "expression~1", cores = 1)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{fullModelFormulaStr}{a formula string specifying the full model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{reducedModelFormulaStr}{a formula string specifying the reduced model in differential expression tests (i.e. likelihood ratio tests) for each gene/feature.}

\item{cores}{the number of cores to be used while testing each gene for differential expression}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
Tests each gene for differential expression as a function of progress through a biological process, or according to other covariates as specified.
}

