% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/expr_models.R
\name{disp_calc_helper}
\alias{disp_calc_helper}
\title{Helper function for parallel dispersion modeling}
\usage{
disp_calc_helper(x, modelFormulaStr, expressionFamily)
}
\description{
Helper function for parallel dispersion modeling
}

