% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{disp_calc_helper}
\alias{disp_calc_helper}
\title{Helper function for parallel dispersion modeling}
\usage{
disp_calc_helper(cds, expressionFamily)
}
\arguments{
\item{x}{A numeric vector of expression values}
}
\description{
Helper function for parallel dispersion modeling
}

