% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{estimate_t}
\alias{estimate_t}
\title{Find the most abundant FPKM estimator of single cell FPKM matrix}
\usage{
estimate_t(fpkm_matrix, return_all = F, fpkm_thresh = 0.1, ...)
}
\arguments{
\item{fpkm_matrix}{a matrix of fpkm for single cell RNA-seq values with each row and column representing genes/isoforms and cells. Row and column names should be included}

\item{return_all}{parameter for the intended return results. If setting TRUE, matrix of dmode as well as max mu and min mu of two gaussian distribution mixture will be returned}

\item{fpkm_thresh}{the threshold for FPKM value used in detecting the most abundant FPKM value}
}
\value{
an vector of most abundant fpkm value corresponding to the transcrpt copy 1. If setting return_all = TRUE, the mode based on gaussian density function and the max or min
mode from the mixture gaussian model
}
\description{
Find the most abundant FPKM estimator of single cell FPKM matrix
}
\details{
This function estimates the most abundant FPKM (t^*) using gaussian kernel density function. It can also optionally output the t^* based on a two gaussian mixture model
based on the smsn.mixture from mixsmsn package
}
\examples{
\dontrun{
HSMM_fpkm_matrix <- exprs(HSMM)
t_estimate = estimate_t(HSMM_fpkm_matrix)
}
}

