% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{estimateDispersionsForCellDataSet}
\alias{estimateDispersionsForCellDataSet}
\title{Helper function to estimate dispersions}
\usage{
estimateDispersionsForCellDataSet(cds, modelFormulaStr, relative_expr, cores)
}
\description{
Helper function to estimate dispersions
}

