% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{estimateDispersionsForCellDataSet}
\alias{estimateDispersionsForCellDataSet}
\title{Helper function to estimate dispersions}
\usage{
estimateDispersionsForCellDataSet(cds, modelFormulaStr, relative_expr,
<<<<<<< HEAD
  min_cells_detected, removeOutliers, cores, verbose = F)
=======
  min_cells_detected, removeOutliers, verbose = FALSE)
>>>>>>> ce5de347ab461f2ed0e6f34051f63b2cef77e50c
}
\arguments{
\item{cds}{a CellDataSet that contains all cells user wants evaluated}

\item{modelFormulaStr}{a formula string specifying the model to fit for the genes.}

\item{relative_expr}{Whether to transform expression into relative values}

\item{min_cells_detected}{Only include genes detected above lowerDetectionLimit in at least this many cells in the dispersion calculation}

\item{removeOutliers}{a boolean it determines whether or not outliers from the data should be removed}

<<<<<<< HEAD
\item{cores}{the number of cores to be used while testing each gene for differential expression.}

\item{verbose}{Whether to show VGAM errors and warnings.}
=======
\item{verbose}{Whether to show detailed running information.}
>>>>>>> ce5de347ab461f2ed0e6f34051f63b2cef77e50c
}
\description{
Helper function to estimate dispersions
}
