% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_cluster_markers}
\alias{find_cluster_markers}
\title{Find marker genes for each group of cells}
\usage{
find_cluster_markers(cds, pr_graph_test_res, group_by = "Cluster",
  qval_threshold = 0.05, morans_I_threshold = 0.25, lower_threshold = 0,
  pseudocount = 1, top_n_by_group = NULL, block_size = 100,
  verbose = FALSE, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{pr_graph_test_res}{the result returned from principalGraphTest}

\item{group_by}{a column in the pData specifying the groups for calculating the specificities. Its default value is Cluster.}

\item{qval_threshold}{The q-value threshold for genes to be selected}

\item{morans_I_threshold}{The lowest Morans' I threshold for selecting genes}

\item{lower_threshold}{The lowest gene expression threshold for genes to be considered as expressed}

\item{pseudocount}{Pseduo-count added to gene expression before calculating the log}

\item{top_n_by_group}{Select top_n_by_group from each group based on the specificity}

\item{block_size}{the number of genes to process per bloclk which dramatically reduces the memory footprints. Default to be 100.}

\item{verbose}{Whether to show VGAM errors and warnings. Only valid for cores = 1.}

\item{...}{Additional arguments passed to function}
}
\value{
a tibble (Tibbles are a modern take on data frame) containing mean expression in each group (mean), percentage of cells 
expressed in each group (percentage), specifity score (specifity) for each gene in each cluster and the principalGraphTest statistics 
(by default, the Moran's I test), together with the information from pData.
}
\description{
Tests each gene for differential expression as a function of pseudotime
or according to other covariates as specified. \code{differentialGeneTest} is
Monocle's main differential analysis routine.
It accepts a CellDataSet and two model formulae as input, which specify generalized
lineage models as implemented by the \code{VGAM} package.
In Monocle 3, we incorporated the find_cluster_markers to find marker genes for each group of cells.
In order to calculate the specificity of each gene to each cluster, we first create a perfect specific
distribution (Q) of each gene by assigning probability 1 to one particular cluster while all 0 to any
other clusters. Then we calculate the percentage of cells expressed in each cluster and convert this percentage
to a probability distribution (P) by dividing the sum of the percentages across all clusters. Then we calculate
the specificity score as 1 - Jensen-Shannon distance between this distribution to the perfect specific distribution.
The Jensen-Shannon divergence is a method measuring the similarity between two probability distributions (P,Q).
The Jensen-Shannon divergence, defined for two distributions P and Q by
\eqn{JSD(P,Q)}{} \eqn{\frac{1}{2} \cdot D(P|M) \cdot \frac{1}{2} \cdot D(Q|M)}{}
where \eqn{M \cdot \frac{1}{2} \cdot P + Q}{} and \eqn{D(A|B)}{} is the Kullback-Leibler divergence.
}
\seealso{
\code{\link{principalGraphTest}}
}
