% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_expression.R
\name{find_cluster_markers}
\alias{find_cluster_markers}
\title{Find marker genes for each group of cells}
\usage{
find_cluster_markers(cds, spatial_res, group_by = "Cluster",
  qval_threshold = 0.05, morans_I_threshold = 0.25, lower_threshold = 0,
  pseudocount = 1, top_n_by_group = NULL, verbose = FALSE, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{spatial_res}{the result returned from spatialDifferentialTest}

\item{group_by}{a column in the pData specifying the groups for calculating the specifities. By default it is Cluster}

\item{qval_threshold}{The q-value threshold for genes to be selected}

\item{morans_I_threshold}{The lowest Morans' I threshold for selecting genes}

\item{lower_threshold}{The lowest gene expression threshold for genes to be considered as expressed}

\item{pseudocount}{Pseduo-count added to gene expression before calculating the log}

\item{top_n_by_group}{Select top_n_by_group from each group based on the specificity}

\item{verbose}{Whether to show VGAM errors and warnings. Only valid for cores = 1.}
}
\value{
a data frame containing the p values and q-values from the likelihood ratio tests on the parallel arrays of models.
}
\description{
Tests each gene for differential expression as a function of pseudotime 
or according to other covariates as specified. \code{differentialGeneTest} is
Monocle's main differential analysis routine. 
It accepts a CellDataSet and two model formulae as input, which specify generalized
lineage models as implemented by the \code{VGAM} package.
}
\seealso{
\code{\link[spatialDifferentialTest]{spatialDifferentialTest}}
}
