% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{fit_model_helper}
\alias{fit_model_helper}
\title{Helper function for parallel VGAM fitting}
\usage{
fit_model_helper(x, modelFormulaStr, expressionFamily, relative_expr,
  disp_func = NULL, pseudocount = 0, verbose = FALSE, weights = NULL,
  ...)
}
\description{
Helper function for parallel VGAM fitting
}

