% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_classic_muscle_markers}
\alias{get_classic_muscle_markers}
\title{Return the names of classic muscle genes}
\usage{
get_classic_muscle_markers()
}
\description{
Returns a list of classic muscle genes. Used to
add conveinence for loading HSMM data.
}
