% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_correct_root_state}
\alias{get_correct_root_state}
\title{A helper function to identify the root principal points}
\usage{
get_correct_root_state(cds, cell_phenotype, root_type)
}
\arguments{
\item{cds}{CellDataSet}

\item{cell_phenotype}{A column in the pData. This describes the characteristics of the cells in the cds}

\item{root_type}{A value in from the cell_phenotype column that corresponds to the predicted starting cell}
}
\description{
A helper function to identify the root principal points
}
