% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{multi_tree_DDRTree}
\alias{multi_tree_DDRTree}
\title{the following functioin is used to learn trajectory on each disjointed components}
\usage{
multi_tree_DDRTree(cds, scale = scale, RGE_method,
  partition_group = "louvain_component", irlba_pca_res, max_components,
  extra_arguments, close_loop = FALSE, verbose = FALSE)
}
\arguments{
\item{cds}{CellDataSet}
}
\description{
the following functioin is used to learn trajectory on each disjointed components
}
