% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{newCellTypeHierarchy}
\alias{newCellTypeHierarchy}
\title{Creates a new CellTypeHierarcy object}
\usage{
newCellTypeHierarchy()
}
\value{
A new CellTypeHierarchy object
}
\description{
Creates a new CellTypeHierarcy object
}

