% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalization.R
\name{optim_mc_func_fix_c}
\alias{optim_mc_func_fix_c}
\title{Calculate the optimization function based on mode of transcript counts, Jessen-Shannon distance as well as the hypothetical total RNA counts}
\usage{
optim_mc_func_fix_c(m, c, t_estimate = estimate_t(TPM_isoform_count_cds),
  relative_expr_matrix = relative_expr_matrix,
  split_relative_expr_matrix = split_relative_exprs, alpha = rep(1,
  ncol(relative_expr_matrix)), total_RNAs = rep(50000,
  ncol(relative_expr_matrix)), cores = 1, weight = 0.5,
  add_kl_divergence = T, verbose = F, ...)
}
\description{
Calculate the optimization function based on mode of transcript counts, Jessen-Shannon distance as well as the hypothetical total RNA counts
}

