% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{order_p_node}
\alias{order_p_node}
\title{Return an ordering for a P node in the PQ tree}
\usage{
order_p_node(q_level_list, dist_matrix)
}
\arguments{
\item{q_level_list}{A list of Q nodes in the PQ tree}

\item{dist_matrix}{A symmetric matrix of pairwise distances between cells}
}
\description{
Return an ordering for a P node in the PQ tree
}
