% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{orderCells}
\alias{orderCells}
\title{Orders cells according to progress through a learned biological process.}
\usage{
orderCells(cds, num_paths = 1, reverse = FALSE, root_cell = NULL,
  scale_pseudotime = F)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{num_paths}{the number of end-point cell states to allow in the biological process.}

\item{reverse}{whether to reverse the beginning and end points of the learned biological process.}

\item{root_cell}{the name of a cell to use as the root of the ordering tree.}

\item{scale_pseudotime}{a logic flag to determine whether or not to scale the pseudotime. If this is set to be true, then the pData dataframe of the returned CDS included an additional column called scale_pseudotime which store the scaled pseudotime values.}
}
\value{
an updated CellDataSet object, in which phenoData contains values for State and Pseudotime for each cell
}
\description{
Orders cells according to progress through a learned biological process.
}

