% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_cluster_graph}
\alias{plot_cluster_graph}
\title{Plots a graph of louvain cluster, layouted with layout_component function.}
\usage{
plot_cluster_graph(cds)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{...}{additional arguments passed into the scale_color_viridis function}
}
\value{
a ggplot2 plot object
}
\description{
Plots a graph of louvain cluster, layouted with layout_component function.
}
\examples{
\dontrun{
library(HSMMSingleCell)
HSMM <- load_HSMM()
HSMM <- reduceD
plot_cell_clusters(HSMM)
plot_cell_clusters(HSMM, color_by="Pseudotime")
plot_cell_clusters(HSMM, markers="MYH3")
}

}
