% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_coexpression_matrix}
\alias{plot_coexpression_matrix}
\title{Plot the branch genes in pseduotime with separate lineage curves
TO DO: add the confidence interval, add p-val, fix error when we cannot generate fitting for the data}
\usage{
plot_coexpression_matrix(cds, rowgenes, colgenes, relative_expr = TRUE,
  min_expr = NULL, cell_size = 0.85, label_by_short_name = TRUE,
  show_density = TRUE)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{rowgenes}{Gene ids or short names to be arrayed on the vertical axis.}

\item{colgenes}{Gene ids or short names to be arrayed on the horizontal axis}

\item{relative_expr}{Whether to transform expression into relative values}

\item{min_expr}{The minimum level of expression to show in the plot}
}
\value{
a ggplot2 plot object
}
\description{
Plot the branch genes in pseduotime with separate lineage curves
TO DO: add the confidence interval, add p-val, fix error when we cannot generate fitting for the data
}

