% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_coexpression_matrix}
\alias{plot_coexpression_matrix}
\title{Not sure we're ready to release this one quite yet:
Plot the branch genes in pseduotime with separate branch curves}
\usage{
plot_coexpression_matrix(
  cds,
  rowgenes,
  colgenes,
  relative_expr = TRUE,
  min_expr = NULL,
  cell_size = 0.85,
  label_by_short_name = TRUE,
  show_density = TRUE,
  round_expr = FALSE
)
}
\arguments{
\item{cds}{CellDataSet for the experiment}

\item{rowgenes}{Gene ids or short names to be arrayed on the vertical axis.}

\item{colgenes}{Gene ids or short names to be arrayed on the horizontal axis}

\item{relative_expr}{Whether to transform expression into relative values}

\item{min_expr}{The minimum level of expression to show in the plot}

\item{cell_size}{A number how large the cells should be in the plot}

\item{label_by_short_name}{a boolean that indicates whether cells should be labeled by their short name}

\item{show_density}{a boolean that indicates whether a 2D density estimation should be shown in the plot}

\item{round_expr}{a boolean that indicates whether cds_expr values should be rounded or not}
}
\value{
a ggplot2 plot object
}
\description{
Not sure we're ready to release this one quite yet:
Plot the branch genes in pseduotime with separate branch curves
}
