% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_gene_pair_spatial_correlation}
\alias{plot_gene_pair_spatial_correlation}
\title{Show the gene pair expression correlation across low dimensional embedding}
\usage{
plot_gene_pair_spatial_correlation(cds, x, y, log = T)
}
\description{
Show the gene pair expression correlation across low dimensional embedding
}
