% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_genes_heatmap}
\alias{plot_genes_heatmap}
\title{Plots a pseudotime-ordered, row-centered heatmap}
\usage{
plot_genes_heatmap(cds, trend_formula = "~sm.ns(Pseudotime, df=3)",
  rescaling = "row", clustering = "row", hclust_method = "ward.D2",
  num_clusters = 6, show_rownames = F, logMode = TRUE,
  pseudocount = 0.1, use_vst = TRUE, border = FALSE, hmcols = NULL,
  scaleMax = 2, scaleMin = -2, return_heatmap = FALSE,
  relative_expr = TRUE, cores = 1, ...)
}
\description{
Plots a pseudotime-ordered, row-centered heatmap
}

