% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_local_spatial_statistics}
\alias{plot_local_spatial_statistics}
\title{plot the local moran's I or the local g for the cells}
\usage{
plot_local_spatial_statistics(cds, markers, method = "local_G", log = TRUE,
  zero.policy = TRUE, return_all = FALSE, ...)
}
\arguments{
\item{cds}{a CellDataSet object upon which to perform this operation}

\item{markers}{a gene name or gene id to use for setting the size of each cell in the plot}

\item{method}{a character string specifying the method (the default 'local_G' or 'local_moran') for detecting local spatial statistics along the principal graph embedded in the low dimensional space.}

\item{log}{a logic value to determine whether or not we need to log the gene expression}

\item{return_all}{A logical argument to determine whether or not the dataframe of the local G or Moran's I should be returned}

\item{color_by}{the cell attribute (e.g. the column of pData(cds)) to map to each cell's color}

\item{quiet}{default NULL, use !verbose global option value; if TRUE, output of summary of influence object suppressed}

\item{labels}{character labels for points with high influence measures, if set to FALSE, no labels are plotted for points with large influence}
}
\description{
plot the local moran's I or the local g for the cells
}
