% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_ordering_genes}
\alias{plot_ordering_genes}
\title{Plots genes by mean vs. dispersion, highlighting those selected for ordering}
\usage{
plot_ordering_genes(cds)
}
\arguments{
\item{cds}{The CellDataSet to be used for the plot.}
}
\description{
Each gray point in the plot is a gene. The black dots are those that were included
in the last call to setOrderingFilter. The red curve shows the mean-variance 
model learning by estimateDispersions().
}
