% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{project_to_representatives}
\alias{project_to_representatives}
\title{Apply embedding smoothing techniques, inlcuding PSL, SSE, force-directed layout, etc. on the kNN graph built from all cells or the downsampled representive 
cells (if number of samples is large than 2000) and then project other non-representative cells to the low dimensional space with five nearest (on original 
UMAP or PCA space) representive cells' coordinates.}
\usage{
project_to_representatives(data, data_ori, landmark_id, louvain_res, pd,
  method = c("PSL", "drl", "fr", "kk", "SSE"), start.temp = NULL,
  k = 20, do_partition = F, max_components = 2, verbose, ...)
}
\arguments{
\item{data}{(dowmsampled) data to perform Louvain clustering and kNN graph construction}

\item{data_ori}{All data points without downsampling. This data space will be used to find the nearest five points for projecting non-landmark points}

\item{landmark_id}{The index of cells}

\item{louvain_res}{The result of louvain clustering from the original space}

\item{pd}{the data frame of phenotype information}

\item{method}{The force directed layout function to use. Only relevant to when drl, fr or kk are used as fdl methods}

\item{start.temp}{argument passed into layout_with_fr function}

\item{do_partition}{Whether or not to separate participation groups and apply FDL in each partition or directly select equal number of representatives in each louvain groups.}

\item{max_components}{the dimensionality of the reduced space}

\item{verbose}{Wheter to print all running details}

\item{...}{additional arguments passed to functions (louvain_clustering) called by this function.}
}
\description{
Apply embedding smoothing techniques, inlcuding PSL, SSE, force-directed layout, etc. on the kNN graph built from all cells or the downsampled representive 
cells (if number of samples is large than 2000) and then project other non-representative cells to the low dimensional space with five nearest (on original 
UMAP or PCA space) representive cells' coordinates.
}
