% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{reduceDimension}
\alias{reduceDimension}
\title{Computes a projection of a CellDataSet object into a lower dimensional space}
\usage{
reduceDimension(cds, max_components = 2, method = c("DDRTree", "ICA"),
  pseudo_expr = NULL, residualModelFormulaStr = NULL, use_vst = NULL,
  verbose = FALSE, use_irlba = NULL, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{max_components}{the dimensionality of the reduced space}

\item{pseudo_expr}{amount to increase expression values before dimensionality reduction}

\item{use_irlba}{Whether to use the IRLBA package for ICA reduction.}

\item{...}{additional arguments to pass to the dimensionality reduction function}

\item{batch}{a vector of labels specifying batch for each cell, the effects of which will be removed prior to dimensionality reduction.}

\item{covariates}{a numeric vector or matrix specifying continuous effects to be removed prior to dimensionality reduction}
}
\value{
an updated CellDataSet object
}
\description{
Computes a projection of a CellDataSet object into a lower dimensional space
}
\details{
Currently, Monocle supports dimensionality reduction with Independent Component Analysis (ICA).
}

