% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/normalization.R
\name{relative2abs}
\alias{relative2abs}
\title{Transform relative expression values into absolute transcript counts}
\usage{
relative2abs(relative_expr_matrix,
  t_estimate = estimate_t(relative_expr_matrix),
  detection_threshold = 0.01430512, ERCC_controls = NULL,
  ERCC_annotation = NULL, volume = 10, dilution = 40000,
  spike_info = NULL, return_all = FALSE, cores = 1, verbose = FALSE)
}
\arguments{
\item{relative_expr_matrix}{an matrix of relative expression values for single cell RNA-seq values with each row and column representing genes/isoforms and cells. Row and column names should be included}

\item{t_estimate}{an vector for the estimated most abundant FPKM value of isoform for a single cell. Estimators based on gene-level relative expression can also give good approximation but estimators
based on isoform FPKM will give better results in general}

\item{detection_threshold}{the lowest ERCC concentration used as a sequencing detection limit, by default is 0.01430512 attomole / Ul. Note that by default we use Mix 1 in ERCC spike-in kit. For all other concentrations, please refer to the illumina ERCC spikein USE GUIDE.}

\item{ERCC_controls}{the FPKM matrix for each ERCC spike-in transcript in the cells if user wants to perform the transformation based on their spike-in data. Note that the row and column names should match up with the ERCC_annotation and relative_exprs_matrix respectively.}

\item{ERCC_annotation}{the ERCC_annotation matrix from illumina USE GUIDE which will be ued for calculating the ERCC transcript copy number for performing the transformation.}

\item{volume}{the approximate volume of the lysis chamber (nanoliters). Default is 10}

\item{dilution}{the dilution of the spikein transcript in the lysis reaction mix. Default is 40, 000The number of spike-in transcripts per single-cell lysis reaction was calculated from}

\item{return_all}{parameter for the intended return results. If setting TRUE, matrix of m, c, k^*, b^* as well as the transformed absolute cds will be returned
in a list format}

\item{verbose:}{a logic flag to determine whether or not we should print all the optimization details}
}
\value{
an matrix of absolute count for isoforms or genes after the transformation. For more details on other output, please refer to detail
}
\description{
Transform a relative expression matrix to absolute transcript matrix based on the decomposed linear regression parameters from most abundant isoform relative expression value.
This function takes a relative expression matrix and a vector of estimated most abundant expression value from the isoform-level matrix and transform it into absolute transcript number.
It is based on the fact that most isoforms of gene in a single cell only express one copy so that the most abundant isoform FPKM (t^*) will corresponding to 1 copy transcript. The
function takes the the vector t^* and then decomposes it into two parameters vectors k^* and b^* (corresponding to each cell) which correspond to the slope and intercept when
we perform the robust linear regression for the spikein data. This decomposition is based on an observed relationship between k and b in terms of b = -3.652201 k + 2.263576. The
function will then apply a linear transformation to convert the FPKM to estimated absolute transcript counts based on the the k^* and b^*. The function can also apply a global
adjustment if setting global_scaling = TRUE. The k*, b* parameters vectors and the global scaling factor can be output in a list format (together with norm_cds) if setting return_all
== TRUE
}
\examples{
\dontrun{
HSMM_relative_expr_matrix <- exprs(HSMM)
HSMM_abs_matrix <- relative2abs(HSMM_relative_expr_matrix, t_estimate = estimate_t(HSMM_relative_expr_matrix))
}
}

