% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_models.R
\name{responseMatrix}
\alias{responseMatrix}
\title{Calculates response values.}
\usage{
responseMatrix(models, newdata = NULL, response_type = "response", cores = 1)
}
\arguments{
\item{models}{a list of models, e.g. as returned by fitModels()}

\item{newdata}{a dataframe used to generate new data for interpolation of time points}

\item{response_type}{the response desired, as accepted by VGAM's predict function}

\item{cores}{number of cores used for calculation}
}
\value{
a matrix where each row is a vector of response values for a particular feature's model, and columns are cells.
}
\description{
Generates a matrix of response values for a set of fitted models
}
