% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/order_cells.R
\name{scale_pseudotime}
\alias{scale_pseudotime}
\title{Scale pseudotime to be in the range from 0 to 100 (it works both for situations involving only one state and complex states)}
\usage{
scale_pseudotime(cds, verbose = F)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{ordering_genes}{a vector of feature ids (from the CellDataSet's featureData) used for ordering cells}
}
\value{
an updated CellDataSet object which an
}
\description{
Scale pseudotime to be in the range from 0 to 100 (it works both for situations involving only one state and complex states)
}

