% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_cells}
\alias{select_cells}
\title{Select cells in an RGL scene}
\usage{
select_cells(cds)
}
\arguments{
\item{cds}{CellDataSet that you'd like to select cells from}
}
\description{
Select cells in an RGL scene
}
