% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{selectCells}
\alias{selectCells}
\title{Select cells in an RGL scene}
\usage{
selectCells(cds, show_pp_only = FALSE)
}
\arguments{
\item{cds}{CellDataSet that you'd like to select cells from}

\item{show_pp_only}{whether to show only the principal points. You should always turn this on for large datasets to avoid computational overloading.}
}
\description{
This function provides an interactive session to allow the users to select multiple cells or region
}
