% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-CellTypeHierarchy.R
\name{selectTopMarkers}
\alias{selectTopMarkers}
\title{Select the most cell type specific markers}
\usage{
selectTopMarkers(marker_specificities, num_markers = 10)
}
\description{
Select the most cell type specific markers
}

