% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{selectTrajectoryRoots}
\alias{selectTrajectoryRoots}
\title{Select the roots of the principal graph}
\usage{
selectTrajectoryRoots(cds, x = 1, y = 2, num_roots = NULL,
  pch = 19, ...)
}
\arguments{
\item{cds}{CellDataSet where roots will be selected from}

\item{x}{The first dimension to plot}

\item{y}{The number of dimension to plot}

\item{num_roots}{Number of roots for the trajectory}

\item{pch}{Size of the principal graph node}

\item{...}{Extra arguments to pass to function}
}
\description{
Select the roots of the principal graph
}
