% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_cells.R
\name{smoothEmbedding}
\alias{smoothEmbedding}
\title{This function tries to learn a smooth embedding from the noisy reduced dimension using different techniques}
\usage{
smoothEmbedding(cds, max_components = 2, smooth_method = c("SSE"),
  verbose = FALSE, ...)
}
\arguments{
\item{cds}{the CellDataSet upon which to perform this operation}

\item{max_components}{the dimensionality of the smoothed reduced space}

\item{smooth_method}{A character string specifying the algorithm to use for smoothing the embedding}

\item{verbose}{Whether to emit verbose output during dimensionality reduction}

\item{...}{additional arguments to pass to the smoothEmbedding function}
}
\value{
an updated CellDataSet object
}
\description{
The function relies on smooth skeleton learning or force direct layout function to learn a smoothier
representation of the data. It can be used to facilitate visualization of the data or downstream graph learning
}
